@@ -1,38 +1,57 @@
-import datetime
-from typing import List
-
-from flask import current_app
-from io import BytesIO
-import pytest
-import random
-from selenium import webdriver
-import time
-import threading
-from xvfbwrapper import Xvfb
-from faker import Faker
 import csv
-import uuid
-import sys
+import datetime
+import math
 import os
+import random
+import sys
+import threading
+import time
+import uuid
+from io import BytesIO
+from pathlib import Path
+from typing import List, Optional
+
+import pytest
+from faker import Faker
+from flask import current_app
 from PIL import Image as Pimage
+from selenium import webdriver
+from sqlalchemy.orm import scoped_session, sessionmaker
+from xvfbwrapper import Xvfb
 
 from OpenOversight.app import create_app, models
-from OpenOversight.app.utils import merge_dicts
-from OpenOversight.app.models import db as _db, Unit, Job, Officer
+from OpenOversight.app.models import Job, Officer, Unit
+from OpenOversight.app.models import db as _db
+from OpenOversight.app.utils.constants import ENCODING_UTF_8
+from OpenOversight.app.utils.general import merge_dicts
 from OpenOversight.tests.routes.route_helpers import ADMIN_EMAIL, ADMIN_PASSWORD
+
 
 factory = Faker()
 
 
-OFFICERS = [('IVANA', '', 'TINKLE'),
-            ('SEYMOUR', '', 'BUTZ'),
-            ('HAYWOOD', 'U', 'CUDDLEME'),
-            ('BEA', '', 'O\'PROBLEM'),
-            ('URA', '', 'SNOTBALL'),
-            ('HUGH', '', 'JASS')]
+OFFICERS = [
+    ("IVANA", "", "TINKLE"),
+    ("SEYMOUR", "", "BUTZ"),
+    ("HAYWOOD", "U", "CUDDLEME"),
+    ("BEA", "", "O'PROBLEM"),
+    ("URA", "", "SNOTBALL"),
+    ("HUGH", "", "JASS"),
+]
 
-RANK_CHOICES_1 = ['Not Sure', 'Police Officer', 'Captain', 'Commander']
-RANK_CHOICES_2 = ['Not Sure', 'Police Officer', 'Lieutenant', 'Sergeant', 'Commander', 'Chief']
+RANK_CHOICES_1 = ["Not Sure", "Police Officer", "Captain", "Commander"]
+RANK_CHOICES_2 = [
+    "Not Sure",
+    "Police Officer",
+    "Lieutenant",
+    "Sergeant",
+    "Commander",
+    "Chief",
+]
+
+DEPARTMENT_NAME = "Springfield Police Department"
+OTHER_DEPARTMENT_NAME = "Chicago Police Department"
+DEPARTMENT_WITHOUT_OFFICERS_NAME = "Empty Police Department"
 
 
 AC_DEPT = 1
@@ -45,26 +64,28 @@ def pick_birth_date():
 def pick_date(seed: bytes = None, start_year=2000, end_year=2020):
     # source: https://stackoverflow.com/questions/40351791/how-to-hash-strings-into-a-float-in-01
     # Wanted to deterministically create a date from a seed string (e.g. the hash or uuid on an officer object)
-    from struct import unpack
     from hashlib import sha256
+    from struct import unpack
 
     def bytes_to_float(b):
-        return float(unpack('L', sha256(b).digest()[:8])[0]) / 2 ** 64
+        return float(unpack("L", sha256(b).digest()[:8])[0]) / 2**64
 
     if seed is None:
-        seed = str(uuid.uuid4()).encode('utf-8')
+        seed = str(uuid.uuid4()).encode(ENCODING_UTF_8)
 
-    return datetime.datetime(start_year, 1, 1, 00, 00, 00) \
-        + datetime.timedelta(days=365 * (end_year - start_year) * bytes_to_float(seed))
+    return datetime.datetime(start_year, 1, 1, 00, 00, 00) + datetime.timedelta(
+        days=365 * (end_year - start_year) * bytes_to_float(seed)
+    )
 
 
 def pick_race():
-    return random.choice(['WHITE', 'BLACK', 'HISPANIC', 'ASIAN',
-                          'PACIFIC ISLANDER', 'Not Sure'])
+    return random.choice(
+        ["WHITE", "BLACK", "HISPANIC", "ASIAN", "PACIFIC ISLANDER", "Not Sure"]
+    )
 
 
 def pick_gender():
-    return random.choice(['M', 'F', 'Other', None])
+    return random.choice(["M", "F", "Other", None])
 
 
 def pick_first():
@@ -100,25 +121,33 @@ def pick_salary():
     return random.randint(100, 100000000) / 100
 
 
-def generate_officer():
+def generate_officer(department):
     year_born = pick_birth_date()
     f_name, m_initial, l_name = pick_name()
     return models.Officer(
-        last_name=l_name, first_name=f_name,
+        last_name=l_name,
+        first_name=f_name,
         middle_initial=m_initial,
-        race=pick_race(), gender=pick_gender(),
+        race=pick_race(),
+        gender=pick_gender(),
         birth_year=year_born,
         employment_date=datetime.datetime(year_born + 20, 4, 4, 1, 1, 1),
-        department_id=pick_department().id,
-        unique_internal_identifier=pick_uid()
+        department_id=department.id,
+        unique_internal_identifier=pick_uid(),
     )
 
 
-def build_assignment(officer: Officer, units: List[Unit], jobs: Job):
-    return models.Assignment(star_no=pick_star(), job_id=random.choice(jobs).id,
-                             officer=officer, unit_id=random.choice(units).id,
-                             star_date=pick_date(officer.full_name().encode('utf-8')),
-                             resign_date=pick_date(officer.full_name().encode('utf-8')))
+def build_assignment(officer: Officer, units: List[Optional[Unit]], jobs: Job):
+    unit = random.choice(units)
+    unit_id = unit.id if unit else None
+    return models.Assignment(
+        star_no=pick_star(),
+        job_id=random.choice(jobs).id,
+        officer=officer,
+        unit_id=unit_id,
+        star_date=pick_date(officer.full_name().encode(ENCODING_UTF_8)),
+        resign_date=pick_date(officer.full_name().encode(ENCODING_UTF_8)),
+    )
 
 
 def build_note(officer, user, content=None):
@@ -130,7 +159,8 @@ def build_note(officer, user, content=None):
         officer_id=officer.id,
         creator_id=user.id,
         date_created=date,
-        date_updated=date)
+        date_updated=date,
+    )
 
 
 def build_description(officer, user, content=None):
@@ -142,7 +172,8 @@ def build_description(officer, user, content=None):
         officer_id=officer.id,
         creator_id=user.id,
         date_created=date,
-        date_updated=date)
+        date_updated=date,
+    )
 
 
 def build_salary(officer):
@@ -151,75 +182,76 @@ def build_salary(officer):
         salary=pick_salary(),
         overtime_pay=pick_salary(),
         year=random.randint(2000, 2019),
-        is_fiscal_year=True if random.randint(0, 1) else False)
+        is_fiscal_year=True if random.randint(0, 1) else False,
+    )
 
 
 def assign_faces(officer, images):
     if random.uniform(0, 1) >= 0.5:
         img_id = random.choice(images).id
-        return models.Face(officer_id=officer.id,
-                           img_id=img_id,
-                           original_image_id=img_id,
-                           featured=False)
+        return models.Face(
+            officer_id=officer.id,
+            img_id=img_id,
+            original_image_id=img_id,
+            featured=False,
+        )
     else:
         return False
 
 
-@pytest.fixture(scope='session')
+@pytest.fixture(scope="session")
 def app(request):
     """Session-wide test `Flask` application."""
-    app = create_app('testing')
-    app.config['WTF_CSRF_ENABLED'] = False
+    app = create_app("testing")
+    app.config["WTF_CSRF_ENABLED"] = False
 
-    # Establish an application context before running the tests.
-    ctx = app.app_context()
-    ctx.push()
-
-    def teardown():
-        ctx.pop()
-
-    request.addfinalizer(teardown)
-    return app
+    yield app
 
 
-@pytest.fixture(scope='session')
-def db(app, request):
+@pytest.fixture(autouse=True)
+def ctx(app):
+    with app.app_context():
+        yield
+
+
+@pytest.fixture(scope="session")
+def db(app):
     """Session-wide test database."""
 
-    def teardown():
-        _db.drop_all()
+    with app.app_context():
+        _db.app = app
+        _db.create_all()
+        connection = _db.engine.connect()
+        session = scoped_session(session_factory=sessionmaker(bind=connection))
+        _db.session = session
+        add_mockdata(session)
+        session.commit()
+        connection.close()
+        session.remove()
 
-    _db.app = app
-    _db.create_all()
-
-    request.addfinalizer(teardown)
-    return _db
+        yield _db
 
 
-@pytest.fixture(scope='function')
-def session(db, request):
+@pytest.fixture(scope="function")
+def session(db):
     """Creates a new database session for a test."""
     connection = db.engine.connect()
     transaction = connection.begin()
 
-    options = dict(bind=connection, binds={})
-    session = db.create_scoped_session(options=options)
-
+    session = scoped_session(session_factory=sessionmaker(bind=connection))
     db.session = session
 
-    def teardown():
-        transaction.rollback()
-        connection.close()
-        session.remove()
+    yield session
 
-    request.addfinalizer(teardown)
-    return session
+    transaction.rollback()
+    connection.close()
+    session.remove()
 
 
 @pytest.fixture
 def test_png_BytesIO():
     test_dir = os.path.dirname(os.path.realpath(__file__))
-    local_path = os.path.join(test_dir, 'images/204Cat.png')
+    local_path = os.path.join(test_dir, "images/204Cat.png")
     img = Pimage.open(local_path)
 
     byte_io = BytesIO()
@@ -231,7 +263,7 @@ def test_png_BytesIO():
 @pytest.fixture
 def test_jpg_BytesIO():
     test_dir = os.path.dirname(os.path.realpath(__file__))
-    local_path = os.path.join(test_dir, 'images/200Cat.jpeg')
+    local_path = os.path.join(test_dir, "images/200Cat.jpeg")
     img = Pimage.open(local_path)
 
     byte_io = BytesIO()
@@ -247,70 +279,101 @@ def test_csv_dir():
 
 
 def add_mockdata(session):
-    NUM_OFFICERS = current_app.config['NUM_OFFICERS']
-    department = models.Department(name='Springfield Police Department',
-                                   short_name='SPD', unique_internal_identifier_label='homer_number')
+    NUM_OFFICERS = current_app.config["NUM_OFFICERS"]
+    assert NUM_OFFICERS >= 5
+    department = models.Department(
+        name=DEPARTMENT_NAME,
+        short_name="SPD",
+        unique_internal_identifier_label="homer_number",
+    )
     session.add(department)
-    department2 = models.Department(name='Chicago Police Department',
-                                    short_name='CPD')
+    department2 = models.Department(name=OTHER_DEPARTMENT_NAME, short_name="CPD")
     session.add(department2)
+    empty_department = models.Department(
+        name=DEPARTMENT_WITHOUT_OFFICERS_NAME, short_name="EPD"
+    )
+    session.add(empty_department)
     session.commit()
 
-    i = 0
-    for rank in RANK_CHOICES_1:
-        session.add(models.Job(
-            job_title=rank,
-            order=i,
-            is_sworn_officer=True,
-            department_id=1
-        ))
-        i += 1
+    for i, rank in enumerate(RANK_CHOICES_1):
+        session.add(
+            models.Job(
+                job_title=rank,
+                order=i,
+                is_sworn_officer=True,
+                department_id=department.id,
+            )
+        )
+        session.add(
+            models.Job(
+                job_title=rank,
+                order=i,
+                is_sworn_officer=True,
+                department_id=empty_department.id,
+            )
+        )
 
-    i = 0
-    for rank in RANK_CHOICES_2:
-        session.add(models.Job(
-            job_title=rank,
-            order=i,
-            is_sworn_officer=True,
-            department_id=2
-        ))
-        i += 1
+    for i, rank in enumerate(RANK_CHOICES_2):
+        session.add(
+            models.Job(
+                job_title=rank,
+                order=i,
+                is_sworn_officer=True,
+                department_id=department2.id,
+            )
+        )
     session.commit()
 
     # Ensure test data is deterministic
-    SEED = current_app.config['SEED']
+    SEED = current_app.config["SEED"]
     random.seed(SEED)
 
     test_units = [
         models.Unit(descrip="test", department_id=1),
-        models.Unit(descrip='District 13', department_id=1),
-        models.Unit(descrip='Donut Devourers', department_id=1),
-        models.Unit(descrip='Bureau of Organized Crime', department_id=2),
-        models.Unit(descrip='Porky\'s BBQ: Rub Division', department_id=2)
+        models.Unit(descrip="District 13", department_id=1),
+        models.Unit(descrip="Donut Devourers", department_id=1),
+        models.Unit(descrip="Bureau of Organized Crime", department_id=2),
+        models.Unit(descrip="Porky's BBQ: Rub Division", department_id=2),
     ]
     session.add_all(test_units)
     session.commit()
+    test_units.append(None)
 
-    test_images = [models.Image(filepath='/static/images/test_cop{}.png'.format(x + 1), department_id=1) for x in range(5)] + \
-        [models.Image(filepath='/static/images/test_cop{}.png'.format(x + 1), department_id=2) for x in range(5)]
+    test_images = [
+        models.Image(
+            filepath=f"/static/images/test_cop{x+1}.png",
+            department_id=department.id,
+        )
+        for x in range(5)
+    ] + [
+        models.Image(
+            filepath=f"/static/images/test_cop{x+1}.png",
+            department_id=department2.id,
+        )
+        for x in range(5)
+    ]
+    session.add_all(test_images)
 
     test_officer_links = [
         models.Link(
-            url='https://openoversight.com/',
-            link_type='link',
-            title='OpenOversight',
-            description='A public, searchable database of law enforcement officers.'),
+            url="https://openoversight.com/",
+            link_type="link",
+            title="OpenOversight",
+            description="A public, searchable database of law enforcement officers.",
+        ),
         models.Link(
-            url='http://www.youtube.com/?v=help',
-            link_type='video',
-            title='Youtube',
-            author='the internet'),
+            url="http://www.youtube.com/?v=help",
+            link_type="video",
+            title="Youtube",
+            author="the internet",
+        ),
     ]
 
-    officers = [generate_officer() for o in range(NUM_OFFICERS)]
+    officers = []
+    for d in [department, department2]:
+        officers += [generate_officer(d) for _ in range(NUM_OFFICERS)]
     officers[0].links = test_officer_links
     session.add_all(officers)
-    session.add_all(test_images)
 
     session.commit()
 
@@ -320,16 +383,31 @@ def add_mockdata(session):
 
     # assures that there are some assigned and unassigned images in each department
     assigned_images_dept1 = models.Image.query.filter_by(department_id=1).limit(3).all()
-    assigned_images_dept2 = models.Image.query.filter_by(department_id=2).limit(2).all()
+    assigned_images_dept2 = models.Image.query.filter_by(department_id=2).limit(3).all()
 
     jobs_dept1 = models.Job.query.filter_by(department_id=1).all()
     jobs_dept2 = models.Job.query.filter_by(department_id=2).all()
-    assignments_dept1 = [build_assignment(officer, test_units, jobs_dept1) for officer in officers_dept1]
-    assignments_dept2 = [build_assignment(officer, test_units, jobs_dept2) for officer in officers_dept2]
+
+    # which percentage of officers have an assignment
+    assignment_ratio = 0.9  # 90%
+    num_officers_with_assignments_1 = math.ceil(len(officers_dept1) * assignment_ratio)
+    assignments_dept1 = [
+        build_assignment(officer, test_units, jobs_dept1)
+        for officer in officers_dept1[:num_officers_with_assignments_1]
+    ]
+    num_officers_with_assignments_2 = math.ceil(len(officers_dept2) * assignment_ratio)
+    assignments_dept2 = [
+        build_assignment(officer, test_units, jobs_dept2)
+        for officer in officers_dept2[:num_officers_with_assignments_2]
+    ]
 
     salaries = [build_salary(officer) for officer in all_officers]
-    faces_dept1 = [assign_faces(officer, assigned_images_dept1) for officer in officers_dept1]
-    faces_dept2 = [assign_faces(officer, assigned_images_dept2) for officer in officers_dept2]
+    faces_dept1 = [
+        assign_faces(officer, assigned_images_dept1) for officer in officers_dept1
+    ]
+    faces_dept2 = [
+        assign_faces(officer, assigned_images_dept2) for officer in officers_dept2
+    ]
     faces1 = [f for f in faces_dept1 if f]
     faces2 = [f for f in faces_dept2 if f]
     session.commit()
@@ -339,64 +417,99 @@ def add_mockdata(session):
     session.add_all(faces1)
     session.add_all(faces2)
 
-    test_user = models.User(email='jen@example.org',
-                            username='test_user',
-                            password='dog',
-                            confirmed=True)
+    test_user = models.User(
+        email="jen@example.org", username="test_user", password="dog", confirmed=True
+    )
     session.add(test_user)
 
-    test_admin = models.User(email=ADMIN_EMAIL,
-                             username='test_admin',
-                             password=ADMIN_PASSWORD,
-                             confirmed=True,
-                             is_administrator=True)
+    test_admin = models.User(
+        email=ADMIN_EMAIL,
+        username="test_admin",
+        password=ADMIN_PASSWORD,
+        confirmed=True,
+        is_administrator=True,
+    )
     session.add(test_admin)
 
-    test_area_coordinator = models.User(email='raq929@example.org',
-                                        username='test_ac',
-                                        password='horse',
-                                        confirmed=True,
-                                        is_area_coordinator=True,
-                                        ac_department_id=AC_DEPT)
+    test_area_coordinator = models.User(
+        email="raq929@example.org",
+        username="test_ac",
+        password="horse",
+        confirmed=True,
+        is_area_coordinator=True,
+        ac_department_id=AC_DEPT,
+    )
     session.add(test_area_coordinator)
 
-    test_unconfirmed_user = models.User(email='freddy@example.org',
-                                        username='b_meson',
-                                        password='dog', confirmed=False)
+    test_unconfirmed_user = models.User(
+        email="freddy@example.org", username="b_meson", password="dog", confirmed=False
+    )
     session.add(test_unconfirmed_user)
+    session.commit()
+
+    test_disabled_user = models.User(
+        email="may@example.org",
+        username="may",
+        password="yam",
+        confirmed=True,
+        is_disabled=True,
+    )
+    session.add(test_disabled_user)
+    session.commit()
+
+    test_modified_disabled_user = models.User(
+        email="sam@example.org",
+        username="sam",
+        password="the yam",
+        confirmed=True,
+        is_disabled=True,
+    )
+    session.add(test_modified_disabled_user)
     session.commit()
 
     test_addresses = [
         models.Location(
-            street_name='Test St',
-            cross_street1='Cross St',
-            cross_street2='2nd St',
-            city='My City',
-            state='AZ',
-            zip_code='23456'),
+            street_name="Test St",
+            cross_street1="Cross St",
+            cross_street2="2nd St",
+            city="My City",
+            state="AZ",
+            zip_code="23456",
+        ),
         models.Location(
-            street_name='Testing St',
-            cross_street1='First St',
-            cross_street2='Fourth St',
-            city='Another City',
-            state='ME',
-            zip_code='23456')
+            street_name="Testing St",
+            cross_street1="First St",
+            cross_street2="Fourth St",
+            city="Another City",
+            state="ME",
+            zip_code="23456",
+        ),
     ]
 
     session.add_all(test_addresses)
     session.commit()
 
     test_license_plates = [
-        models.LicensePlate(number='603EEE', state='MA'),
-        models.LicensePlate(number='404301', state='WA')
+        models.LicensePlate(number="603EEE", state="MA"),
+        models.LicensePlate(number="404301", state="WA"),
     ]
 
     session.add_all(test_license_plates)
     session.commit()
 
     test_incident_links = [
-        models.Link(url='https://stackoverflow.com/', link_type='link', creator=test_admin, creator_id=test_admin.id),
-        models.Link(url='http://www.youtube.com/?v=help', link_type='video', creator=test_admin, creator_id=test_admin.id)
+        models.Link(
+            url="https://stackoverflow.com/",
+            link_type="link",
+            creator=test_admin,
+            creator_id=test_admin.id,
+        ),
+        models.Link(
+            url="http://www.youtube.com/?v=help",
+            link_type="video",
+            creator=test_admin,
+            creator_id=test_admin.id,
+        ),
     ]
 
     session.add_all(test_incident_links)
@@ -406,40 +519,42 @@ def add_mockdata(session):
         models.Incident(
             date=datetime.date(2016, 3, 16),
             time=datetime.time(4, 20),
-            report_number='42',
-            description='### A thing happened\n **Markup** description',
+            report_number="42",
+            description="### A thing happened\n **Markup** description",
             department_id=1,
             address=test_addresses[0],
             license_plates=test_license_plates,
             links=test_incident_links,
             officers=[all_officers[o] for o in range(4)],
             creator_id=1,
-            last_updated_id=1
+            last_updated_id=1,
         ),
         models.Incident(
             date=datetime.date(2017, 12, 11),
             time=datetime.time(2, 40),
-            report_number='38',
-            description='A thing happened',
+            report_number="38",
+            description="A thing happened",
             department_id=2,
             address=test_addresses[1],
             license_plates=[test_license_plates[0]],
             links=test_incident_links,
             officers=[all_officers[o] for o in range(3)],
             creator_id=2,
-            last_updated_id=1
+            last_updated_id=1,
         ),
         models.Incident(
             date=datetime.datetime(2019, 1, 15),
-            report_number='39',
-            description='A test description that has over 300 chars. The purpose is to see how to display a larger descrption. Descriptions can get lengthy. So lengthy. It is a description with a lot to say. Descriptions can get lengthy. So lengthy. It is a description with a lot to say. Descriptions can get lengthy. So lengthy. It is a description with a lot to say. Lengthy lengthy lengthy.',
+            report_number="39",
+            description=(
+                Path(__file__).parent / "description_overflow.txt"
+            ).read_text(),
             department_id=2,
             address=test_addresses[1],
             license_plates=[test_license_plates[0]],
             links=test_incident_links,
             officers=[all_officers[o] for o in range(1)],
             creator_id=2,
-            last_updated_id=1
+            last_updated_id=1,
         ),
     ]
     session.add_all(test_incidents)
@@ -449,7 +564,9 @@ def add_mockdata(session):
 
     # for testing routes
     first_officer = models.Officer.query.get(1)
-    note = build_note(first_officer, test_admin, "### A markdown note\nA **test** note!")
+    note = build_note(
+        first_officer, test_admin, "### A markdown note\nA **test** note!"
+    )
     session.add(note)
     for officer in models.Officer.query.limit(20):
         user = random.choice(users_that_can_create_notes)
@@ -462,7 +579,9 @@ def add_mockdata(session):
 
     # for testing routes
     first_officer = models.Officer.query.get(1)
-    description = build_description(first_officer, test_admin, "### A markdown description\nA **test** description!")
+    description = build_description(
+        first_officer, test_admin, "### A markdown description\nA **test** description!"
+    )
     session.add(description)
     for officer in models.Officer.query.limit(20):
         user = random.choice(users_that_can_create_descriptions)
@@ -476,35 +595,29 @@ def add_mockdata(session):
 
 @pytest.fixture
 def mockdata(session):
-    return add_mockdata(session)
+    pass
 
 
 @pytest.fixture
 def department(session):
-    department = models.Department(
-        id=1,
-        name="Springfield Police Department",
-        short_name="SPD",
-        unique_internal_identifier_label="homer_number",
-    )
-    session.add(department)
-    session.commit()
-    return department
+    return models.Department.query.filter_by(name=DEPARTMENT_NAME).one()
 
 
 @pytest.fixture
-def department_with_ranks(department, session):
-    for order, rank in enumerate(RANK_CHOICES_1):
-        session.add(
-            models.Job(
-                job_title=rank,
-                order=order,
-                is_sworn_officer=True,
-                department=department,
-            )
-        )
-    session.commit()
-    return department
+def department_without_officers(session):
+    return models.Department.query.filter_by(
+        name=DEPARTMENT_WITHOUT_OFFICERS_NAME
+    ).one()
+
+
+@pytest.fixture
+def officer_no_assignments(department):
+    return (
+        Officer.query.filter_by(department_id=department.id)
+        .outerjoin(Officer.assignments)
+        .filter(Officer.assignments == None)  # noqa: E711
+        .first()
+    )
 
 
 @pytest.fixture
@@ -530,60 +643,66 @@ def csvfile(mockdata, tmp_path, request):
     tmp_path.mkdir()
 
     fieldnames = [
-        'department_id',
-        'unique_internal_identifier',
-        'first_name',
-        'last_name',
-        'middle_initial',
-        'suffix',
-        'gender',
-        'race',
-        'employment_date',
-        'birth_year',
-        'star_no',
-        'job_title',
-        'unit_id',
-        'star_date',
-        'resign_date',
-        'salary',
-        'salary_year',
-        'salary_is_fiscal_year',
-        'overtime_pay'
+        "department_id",
+        "unique_internal_identifier",
+        "first_name",
+        "last_name",
+        "middle_initial",
+        "suffix",
+        "gender",
+        "race",
+        "employment_date",
+        "birth_year",
+        "star_no",
+        "job_title",
+        "unit_id",
+        "star_date",
+        "resign_date",
+        "salary",
+        "salary_year",
+        "salary_is_fiscal_year",
+        "overtime_pay",
     ]
 
     officers_dept1 = models.Officer.query.filter_by(department_id=1).all()
 
     if sys.version_info.major == 2:
-        csvf = open(str(csv_path), 'w')
+        csvf = open(str(csv_path), "w")
     else:
-        csvf = open(str(csv_path), 'w', newline='')
+        csvf = open(str(csv_path), "w", newline="")
     try:
-        writer = csv.DictWriter(csvf, fieldnames=fieldnames, extrasaction='ignore')
+        writer = csv.DictWriter(csvf, fieldnames=fieldnames, extrasaction="ignore")
         writer.writeheader()
         for officer in officers_dept1:
             if not officer.unique_internal_identifier:
                 officer.unique_internal_identifier = str(uuid.uuid4())
-            towrite = merge_dicts(vars(officer), {'department_id': 1})
+            towrite = merge_dicts(vars(officer), {"department_id": 1})
             if len(list(officer.assignments)) > 0:
                 assignment = officer.assignments[0]
-                towrite = merge_dicts(towrite, {
-                    'star_no': assignment.star_no,
-                    'job_title': assignment.job.job_title if assignment.job else None,
-                    'unit_id': assignment.unit_id,
-                    'star_date': assignment.star_date,
-                    'resign_date': assignment.resign_date
-                })
+                towrite = merge_dicts(
+                    towrite,
+                    {
+                        "star_no": assignment.star_no,
+                        "job_title": assignment.job.job_title
+                        if assignment.job
+                        else None,
+                        "unit_id": assignment.unit_id,
+                        "star_date": assignment.star_date,
+                        "resign_date": assignment.resign_date,
+                    },
+                )
             if len(list(officer.salaries)) > 0:
                 salary = officer.salaries[0]
-                towrite = merge_dicts(towrite, {
-                    'salary': salary.salary,
-                    'salary_year': salary.year,
-                    'salary_is_fiscal_year': salary.is_fiscal_year,
-                    'overtime_pay': salary.overtime_pay
-                })
+                towrite = merge_dicts(
+                    towrite,
+                    {
+                        "salary": salary.salary,
+                        "salary_year": salary.year,
+                        "salary_is_fiscal_year": salary.is_fiscal_year,
+                        "overtime_pay": salary.overtime_pay,
+                    },
+                )
             writer.writerow(towrite)
-    except:  # noqa E722
-        raise
     finally:
         csvf.close()
 
@@ -592,33 +711,42 @@ def csvfile(mockdata, tmp_path, request):
 
 
 @pytest.fixture
-def client(app, request):
-    client = app.test_client()
-
-    def teardown():
-        pass
-    request.addfinalizer(teardown)
-    return client
+def client(app):
+    with app.app_context():
+        client = app.test_client()
+        yield client
 
 
-@pytest.fixture
-def browser(app, request):
+@pytest.fixture(scope="session")
+def worker_number(worker_id):
+    if len(worker_id) < 2 or worker_id[:2] != "gw":
+        return 0
+    return int(worker_id[2:])
+
+
+@pytest.fixture(scope="session")
+def server_port(worker_number):
+    return 5000 + worker_number
+
+
+@pytest.fixture(scope="session")
+def browser(app, server_port):
     # start server
-    threading.Thread(target=app.run).start()
+    port = server_port
+    print("Starting server at port {port}")
+    threading.Thread(
+        target=app.run, daemon=True, kwargs={"debug": False, "port": port}
+    ).start()
     # give the server a few seconds to ensure it is up
     time.sleep(10)
 
     # start headless webdriver
     vdisplay = Xvfb()
     vdisplay.start()
-    driver = webdriver.Firefox(log_path='/tmp/geckodriver.log')
+    driver = webdriver.Firefox(log_path="/tmp/geckodriver.log")
     # wait for browser to start up
     time.sleep(3)
     yield driver
-
-    # shutdown server
-    driver.get("http://localhost:5000/shutdown")
-    time.sleep(3)
 
     # shutdown headless webdriver
     driver.quit()
