@@ -1,27 +1,30 @@
-from __future__ import division
+import os
 from contextlib import contextmanager
+
 import pytest
 from selenium.common.exceptions import TimeoutException
-from selenium.webdriver.support.ui import WebDriverWait
+from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions
 from selenium.webdriver.support.select import Select
-from selenium.webdriver.common.by import By
+from selenium.webdriver.support.ui import WebDriverWait
 from sqlalchemy.sql.expression import func
-from OpenOversight.app.models import db, Officer, Incident, Department, Unit
+
 from OpenOversight.app.config import BaseConfig
+from OpenOversight.app.models import Department, Incident, Officer, Unit, db
+
+
+DESCRIPTION_CUTOFF = 700
 
 
 @contextmanager
 def wait_for_page_load(browser, timeout=10):
-    old_page = browser.find_element_by_tag_name('html')
+    old_page = browser.find_element_by_tag_name("html")
     yield
-    WebDriverWait(browser, timeout).until(
-        expected_conditions.staleness_of(old_page)
-    )
+    WebDriverWait(browser, timeout).until(expected_conditions.staleness_of(old_page))
 
 
-def login_admin(browser):
-    browser.get("http://localhost:5000/auth/login")
+def login_admin(browser, server_port):
+    browser.get(f"http://localhost:{server_port}/auth/login")
     with wait_for_page_load(browser):
         elem = browser.find_element_by_id("email")
         elem.clear()
@@ -44,8 +47,18 @@ def wait_for_element(browser, locator, text, timeout=10):
         pytest.fail("Timed out while waiting for element to appear")
 
 
-def test_user_can_load_homepage_and_get_to_form(mockdata, browser):
-    browser.get("http://localhost:5000")
+def wait_for_element_to_be_visible(browser, locator, text, timeout=10):
+    try:
+        element_visible = expected_conditions.visibility_of_element_located(
+            (locator, text)
+        )
+        WebDriverWait(browser, timeout).until(element_visible)
+    except TimeoutException:
+        pytest.fail("Timed out while waiting for element to become visible")
+
+
+def test_user_can_load_homepage_and_get_to_form(mockdata, browser, server_port):
+    browser.get(f"http://localhost:{server_port}")
 
     # Complainant loads homepage
     assert "OpenOversight" in browser.title
@@ -56,8 +69,8 @@ def test_user_can_load_homepage_and_get_to_form(mockdata, browser):
     assert "Find an Officer" in page_text
 
 
-def test_user_can_use_form_to_get_to_browse(mockdata, browser):
-    browser.get("http://localhost:5000/find")
+def test_user_can_use_form_to_get_to_browse(mockdata, browser, server_port):
+    browser.get(f"http://localhost:{server_port}/find")
 
     # Complainant selects department and proceeds to next step
     browser.find_element_by_id("activate-step-2").click()
@@ -79,8 +92,12 @@ def test_user_can_use_form_to_get_to_browse(mockdata, browser):
     assert browser.find_element_by_id("officer-profile-1")
 
 
-def test_user_can_get_to_complaint(mockdata, browser):
-    browser.get("http://localhost:5000/complaint?officer_star=6265&officer_first_name=IVANA&officer_last_name=SNOTBALL&officer_middle_initial=&officer_image=static%2Fimages%2Ftest_cop2.png")
+def test_user_can_get_to_complaint(mockdata, browser, server_port):
+    browser.get(
+        f"http://localhost:{server_port}/complaint?officer_star=6265&"
+        "officer_first_name=IVANA&officer_last_name=SNOTBALL&officer_middle_initial="
+        "&officer_image=static%2Fimages%2Ftest_cop2.png"
+    )
 
     wait_for_element(browser, By.TAG_NAME, "h1")
     # Complainant arrives at page with the badge number, name, and link
@@ -90,34 +107,39 @@ def test_user_can_get_to_complaint(mockdata, browser):
     assert "File a Complaint" in title_text
 
 
-def test_officer_browse_pagination(mockdata, browser):
+def test_officer_browse_pagination(mockdata, browser, server_port):
     dept_id = 1
     total = Officer.query.filter_by(department_id=dept_id).count()
     perpage = BaseConfig.OFFICERS_PER_PAGE
 
     # first page of results
-    browser.get("http://localhost:5000/department/{}?page=1"
-                .format(dept_id))
+    browser.get(f"http://localhost:{server_port}/department/{dept_id}?page=1")
     wait_for_element(browser, By.TAG_NAME, "body")
     page_text = browser.find_element_by_tag_name("body").text
-    expected = 'Showing 1-{} of {}'.format(perpage, total)
+    expected = "Showing 1-{} of {}".format(perpage, total)
     assert expected in page_text
 
     # last page of results
     last_page_index = (total // perpage) + 1
-    browser.get("http://localhost:5000/department/{}?page={}"
-                .format(dept_id, last_page_index))
+    browser.get(
+        f"http://localhost:{server_port}/department/{dept_id}?page={last_page_index}"
+    )
     wait_for_element(browser, By.TAG_NAME, "body")
     page_text = browser.find_element_by_tag_name("body").text
-    expected = ('Showing {}-{} of {}'
-                .format(perpage * (total // perpage) + 1, total, total))
+    expected = "Showing {}-{} of {}".format(
+        perpage * (total // perpage) + 1, total, total
+    )
     assert expected in page_text
 
 
-def test_find_officer_can_see_uii_question_for_depts_with_uiis(mockdata, browser):
-    browser.get("http://localhost:5000/find")
+def test_find_officer_can_see_uii_question_for_depts_with_uiis(
+    mockdata, browser, server_port
+):
+    browser.get(f"http://localhost:{server_port}/find")
 
-    dept_with_uii = Department.query.filter(Department.unique_internal_identifier_label.isnot(None)).first()
+    dept_with_uii = Department.query.filter(
+        Department.unique_internal_identifier_label.isnot(None)
+    ).first()
     dept_id = str(dept_with_uii.id)
 
     dept_selector = Select(browser.find_element_by_id("dept"))
@@ -128,10 +150,14 @@ def test_find_officer_can_see_uii_question_for_depts_with_uiis(mockdata, browser
     assert "Do you know any part of the Officer's" in page_text
 
 
-def test_find_officer_cannot_see_uii_question_for_depts_without_uiis(mockdata, browser):
-    browser.get("http://localhost:5000/find")
+def test_find_officer_cannot_see_uii_question_for_depts_without_uiis(
+    mockdata, browser, server_port
+):
+    browser.get(f"http://localhost:{server_port}/find")
 
-    dept_without_uii = Department.query.filter_by(unique_internal_identifier_label=None).one_or_none()
+    dept_without_uii = Department.query.filter_by(
+        unique_internal_identifier_label=None
+    ).first()
     dept_id = str(dept_without_uii.id)
 
     dept_selector = Select(browser.find_element_by_id("dept"))
@@ -142,47 +168,79 @@ def test_find_officer_cannot_see_uii_question_for_depts_without_uiis(mockdata, b
     assert len(results) == 0
 
 
-def test_incident_detail_display_read_more_button_for_descriptions_over_300_chars(mockdata, browser):
+def test_incident_detail_display_read_more_button_for_descriptions_over_cutoff(
+    mockdata, browser, server_port
+):
     # Navigate to profile page for officer with short and long incident descriptions
-    browser.get("http://localhost:5000/officer/1")
+    browser.get(f"http://localhost:{server_port}/officer/1")
 
-    incident_long_descrip = Incident.query.filter(func.length(Incident.description) > 300).one_or_none()
+    incident_long_descrip = Incident.query.filter(
+        func.length(Incident.description) > DESCRIPTION_CUTOFF
+    ).one_or_none()
     incident_id = str(incident_long_descrip.id)
 
     result = browser.find_element_by_id("description-overflow-row_" + incident_id)
     assert result.is_displayed()
 
 
-def test_incident_detail_do_not_display_read_more_button_for_descriptions_under_300_chars(mockdata, browser):
+def test_incident_detail_truncate_description_for_descriptions_over_cutoff(
+    mockdata, browser, server_port
+):
     # Navigate to profile page for officer with short and long incident descriptions
-    browser.get("http://localhost:5000/officer/1")
+    browser.get(f"http://localhost:{server_port}/officer/1")
 
-    # Select incident for officer that has description under 300 chars
+    incident_long_descrip = Incident.query.filter(
+        func.length(Incident.description) > DESCRIPTION_CUTOFF
+    ).one_or_none()
+    incident_id = str(incident_long_descrip.id)
+
+    # Check that the text is truncated and contains more than just the ellipsis
+    truncated_text = browser.find_element(
+        "id", "incident-description_" + incident_id
+    ).text
+    assert "…" in truncated_text
+    # Include buffer for jinja rendered spaces
+    assert DESCRIPTION_CUTOFF + 20 > len(truncated_text) > 100
+
+
+def test_incident_detail_do_not_display_read_more_button_for_descriptions_under_cutoff(
+    mockdata, browser, server_port
+):
+    # Navigate to profile page for officer with short and long incident descriptions
+    browser.get(f"http://localhost:{server_port}/officer/1")
+
+    # Select incident for officer that has description under cuttoff chars
     result = browser.find_element_by_id("description-overflow-row_1")
     assert not result.is_displayed()
 
 
-def test_click_to_read_more_displays_full_description(mockdata, browser):
+def test_click_to_read_more_displays_full_description(mockdata, browser, server_port):
     # Navigate to profile page for officer with short and long incident descriptions
-    browser.get("http://localhost:5000/officer/1")
+    browser.get(f"http://localhost:{server_port}/officer/1")
 
-    incident_long_descrip = Incident.query.filter(func.length(Incident.description) > 300).one_or_none()
-    orig_descrip = incident_long_descrip.description
+    incident_long_descrip = Incident.query.filter(
+        func.length(Incident.description) > DESCRIPTION_CUTOFF
+    ).one_or_none()
+    orig_descrip = incident_long_descrip.description.strip()
     incident_id = str(incident_long_descrip.id)
 
     button = browser.find_element_by_id("description-overflow-button_" + incident_id)
     button.click()
 
-    description_text = browser.find_element_by_id("incident-description_" + incident_id).text
+    description_text = browser.find_element_by_id(
+        "incident-description_" + incident_id
+    ).text.strip()
     assert len(description_text) == len(orig_descrip)
     assert description_text == orig_descrip
 
 
-def test_click_to_read_more_hides_the_read_more_button(mockdata, browser):
+def test_click_to_read_more_hides_the_read_more_button(mockdata, browser, server_port):
     # Navigate to profile page for officer with short and long incident descriptions
-    browser.get("http://localhost:5000/officer/1")
+    browser.get(f"http://localhost:{server_port}/officer/1")
 
-    incident_long_descrip = Incident.query.filter(func.length(Incident.description) > 300).one_or_none()
+    incident_long_descrip = Incident.query.filter(
+        func.length(Incident.description) > DESCRIPTION_CUTOFF
+    ).one_or_none()
     incident_id = str(incident_long_descrip.id)
 
     button = browser.find_element_by_id("description-overflow-button_" + incident_id)
@@ -192,46 +250,150 @@ def test_click_to_read_more_hides_the_read_more_button(mockdata, browser):
     assert not buttonRow.is_displayed()
 
 
-def test_officer_form_has_units_alpha_sorted(mockdata, browser):
-    login_admin(browser)
+def test_officer_form_has_units_alpha_sorted(mockdata, browser, server_port):
+    login_admin(browser, server_port)
 
     # get the units from the DB in the sort we expect
-    db_units_sorted = list(map(lambda x: x.descrip, db.session.query(Unit).order_by(Unit.descrip.asc()).all()))
+    db_units_sorted = list(
+        map(
+            lambda x: x.descrip,
+            db.session.query(Unit).order_by(Unit.descrip.asc()).all(),
+        )
+    )
     # the Select tag in the interface has a 'None' value at the start
-    db_units_sorted.insert(0, 'None')
+    db_units_sorted.insert(0, "None")
 
     # Check for the Unit sort on the 'add officer' form
-    browser.get("http://localhost:5000/officer/new")
+    browser.get(f"http://localhost:{server_port}/officer/new")
     unit_select = Select(browser.find_element_by_id("unit"))
     select_units_sorted = list(map(lambda x: x.text, unit_select.options))
     assert db_units_sorted == select_units_sorted
 
     # Check for the Unit sort on the 'add assignment' form
-    browser.get("http://localhost:5000/officer/1")
+    browser.get(f"http://localhost:{server_port}/officer/1")
     unit_select = Select(browser.find_element_by_id("unit"))
     select_units_sorted = list(map(lambda x: x.text, unit_select.options))
     assert db_units_sorted == select_units_sorted
 
 
-def test_edit_officer_form_coerces_none_race_or_gender_to_not_sure(mockdata, browser):
+def test_edit_officer_form_coerces_none_race_or_gender_to_not_sure(
+    mockdata, browser, server_port
+):
     # Set NULL race and gender for officer 1
     db.session.execute(
-        Officer.__table__.update().where(Officer.id == 1).values(race=None, gender=None))
+        Officer.__table__.update().where(Officer.id == 1).values(race=None, gender=None)
+    )
     db.session.commit()
 
-    login_admin(browser)
+    login_admin(browser, server_port)
 
     # Nagivate to edit officer page for officer having NULL race and gender
-    browser.get("http://localhost:5000/officer/1/edit")
+    browser.get(f"http://localhost:{server_port}/officer/1/edit")
 
     wait_for_element(browser, By.ID, "gender")
     select = Select(browser.find_element_by_id("gender"))
     selected_option = select.first_selected_option
     selected_text = selected_option.text
-    assert selected_text == 'Not Sure'
+    assert selected_text == "Not Sure"
 
     wait_for_element(browser, By.ID, "race")
     select = Select(browser.find_element_by_id("race"))
     selected_option = select.first_selected_option
     selected_text = selected_option.text
-    assert selected_text == 'Not Sure'
+    assert selected_text == "Not Sure"
+
+
+@pytest.mark.skip("Enable once real file upload in tests is supported.")
+def test_image_classification_and_tagging(mockdata, browser, server_port):
+    test_dir = os.path.dirname(os.path.realpath(__file__))
+    img_path = os.path.join(test_dir, "images/200Cat.jpeg")
+
+    login_admin(browser, server_port)
+
+    # 1. Create new department (to avoid mockdata)
+    browser.get(f"http://localhost:{server_port}/department/new")
+    wait_for_page_load(browser)
+    browser.find_element(By.ID, "name").send_keys("Auburn Police Department")
+    browser.find_element(By.ID, "short_name").send_keys("APD")
+    browser.find_element(By.ID, "submit").click()
+    wait_for_page_load(browser)
+
+    # 2. Add a new officer
+    browser.get(f"http://localhost:{server_port}/officer/new")
+    wait_for_page_load(browser)
+
+    dept_select = Select(browser.find_element("id", "department"))
+    dept_select.select_by_visible_text("Auburn Police Department")
+    dept_id = dept_select.first_selected_option.get_attribute("value")
+
+    browser.find_element(By.ID, "first_name").send_keys("Officer")
+    browser.find_element(By.ID, "last_name").send_keys("Friendly")
+    browser.find_element(By.ID, "submit").click()
+
+    wait_for_page_load(browser)
+    # expected url: http://localhost:{server_port}/submit_officer_images/officer/<id>
+    officer_id = browser.current_url.split("/")[-1]
+
+    # 3. Submit an image
+    browser.get(f"http://localhost:{server_port}/submit")
+    wait_for_page_load(browser)
+
+    Select(browser.find_element("id", "department")).select_by_value(dept_id)
+
+    # Submit files in selenium: https://stackoverflow.com/a/61566075
+    wait_for_element(browser, By.CLASS_NAME, "dz-hidden-input")
+    upload = browser.find_element(By.CLASS_NAME, "dz-hidden-input")
+    upload.send_keys(img_path)
+    wait_for_element(browser, By.CLASS_NAME, "dz-success")
+
+    # 4. Classify the uploaded image
+    browser.get(f"http://localhost:{server_port}/sort/department/{dept_id}")
+
+    # Check that image loaded correctly: https://stackoverflow.com/a/36296478
+    wait_for_element(browser, By.TAG_NAME, "img")
+    image = browser.find_element(By.TAG_NAME, "img")
+    assert image.get_attribute("complete") == "true"
+    assert int(image.get_attribute("naturalHeight")) > 0
+
+    browser.find_element(By.ID, "answer-yes").click()
+
+    wait_for_page_load(browser)
+    page_text = browser.find_element(By.TAG_NAME, "body").text
+    assert "All images have been classfied!" in page_text
+
+    # 5. Identify the new officer in the uploaded image
+    browser.get(f"http://localhost:{server_port}/cop_face/department/{dept_id}")
+    wait_for_page_load(browser)
+    browser.find_element(By.ID, "officer_id").send_keys(officer_id)
+    browser.find_element(By.CSS_SELECTOR, "input[value='Add identified face']").click()
+
+    wait_for_page_load(browser)
+    page_text = browser.find_element(By.TAG_NAME, "body").text
+    assert "Tag added to database" in page_text
+
+    # 6. Log out as admin
+    browser.get(f"http://localhost:{server_port}/auth/logout")
+    wait_for_page_load(browser)
+
+    # 7. Check that the tag appears on the officer page
+    browser.get(f"http://localhost:{server_port}/officer/{officer_id}")
+    wait_for_page_load(browser)
+    browser.find_element(By.CSS_SELECTOR, "a > img.officer-face").click()
+
+    wait_for_page_load(browser)
+    image = browser.find_element(By.TAG_NAME, "img")
+    frame = browser.find_element(By.ID, "face-tag-frame")
+
+    # 8. Check that the tag frame is fully contained within the image
+    wait_for_element_to_be_visible(browser, By.ID, "face-tag-frame")
+    assert image.location["x"] <= frame.location["x"]
+    assert image.location["y"] <= frame.location["y"]
+    assert (
+        image.location["x"] + image.size["width"]
+        >= frame.location["x"] + frame.size["width"]
+    )
+    assert (
+        image.location["y"] + image.size["height"]
+        >= frame.location["y"] + frame.size["height"]
+    )
+    assert image.location["y"] <= frame.location["y"]
