@@ -1,5 +1,6 @@
 from flask import Blueprint
 
-auth = Blueprint('auth', __name__)  # noqa
 
-from . import views  # noqa
+auth = Blueprint("auth", __name__)
+
+from . import views  # noqa: E402,F401
