@@ -11,40 +11,42 @@ class BaseConfig(object):
     SQLALCHEMY_TRACK_MODIFICATIONS = False
 
     # pagination
-    OFFICERS_PER_PAGE = os.environ.get('OFFICERS_PER_PAGE', 20)
-    USERS_PER_PAGE = os.environ.get('USERS_PER_PAGE', 20)
+    OFFICERS_PER_PAGE = os.environ.get("OFFICERS_PER_PAGE", 20)
+    USERS_PER_PAGE = os.environ.get("USERS_PER_PAGE", 20)
 
     # Form Settings
     WTF_CSRF_ENABLED = True
-    SECRET_KEY = os.environ.get('SECRET_KEY', 'changemeplzorelsehax')
+    SECRET_KEY = os.environ.get("SECRET_KEY", "changemeplzorelsehax")
 
     # Mail Settings
-    MAIL_SERVER = os.environ.get('MAIL_SERVER', 'smtp.googlemail.com')
+    MAIL_SERVER = os.environ.get("MAIL_SERVER", "smtp.googlemail.com")
     MAIL_PORT = 587
     MAIL_USE_TLS = True
-    MAIL_USERNAME = os.environ.get('MAIL_USERNAME')
-    MAIL_PASSWORD = os.environ.get('MAIL_PASSWORD')
-    OO_MAIL_SUBJECT_PREFIX = os.environ.get('OO_MAIL_SUBJECT_PREFIX', '[OpenOversight]')
-    OO_MAIL_SENDER = os.environ.get('OO_MAIL_SENDER', 'OpenOversight <OpenOversight@gmail.com>')
-    # OO_ADMIN = os.environ.get('OO_ADMIN')
+    MAIL_USERNAME = os.environ.get("MAIL_USERNAME")
+    MAIL_PASSWORD = os.environ.get("MAIL_PASSWORD")
+    OO_MAIL_SUBJECT_PREFIX = os.environ.get("OO_MAIL_SUBJECT_PREFIX", "[OpenOversight]")
+    OO_MAIL_SENDER = os.environ.get(
+        "OO_MAIL_SENDER", "OpenOversight <OpenOversight@gmail.com>"
+    )
+    # OO_ADMIN = os.environ.get("OO_ADMIN")
 
     # AWS Settings
-    AWS_ACCESS_KEY_ID = os.environ.get('AWS_ACCESS_KEY_ID')
-    AWS_SECRET_ACCESS_KEY = os.environ.get('AWS_SECRET_ACCESS_KEY')
-    AWS_DEFAULT_REGION = os.environ.get('AWS_DEFAULT_REGION')
-    S3_BUCKET_NAME = os.environ.get('S3_BUCKET_NAME')
-    FACE_ACCESS_KEY_ID = os.environ.get('FACE_ACCESS_KEY_ID')
-    FACE_SECRET_ACCESS_KEY = os.environ.get('FACE_SECRET_ACCESS_KEY')
+    AWS_ACCESS_KEY_ID = os.environ.get("AWS_ACCESS_KEY_ID")
+    AWS_SECRET_ACCESS_KEY = os.environ.get("AWS_SECRET_ACCESS_KEY")
+    AWS_DEFAULT_REGION = os.environ.get("AWS_DEFAULT_REGION")
+    S3_BUCKET_NAME = os.environ.get("S3_BUCKET_NAME")
+    FACE_ACCESS_KEY_ID = os.environ.get("FACE_ACCESS_KEY_ID")
+    FACE_SECRET_ACCESS_KEY = os.environ.get("FACE_SECRET_ACCESS_KEY")
 
 
     # Upload Settings
     MAX_CONTENT_LENGTH = 500 * 1024 * 1024
-    ALLOWED_EXTENSIONS = set(['jpeg', 'jpg', 'jpe', 'png', 'gif', 'webp'])
-    ALLOWED_DOC_EXTENSIONS = set(['jpeg', 'jpg', 'jpe', 'png', 'gif', 'webp',
-        'pdf', 'doc', 'docx', 'xls', 'xlxs', 'ppt', 'pptx',
+    ALLOWED_EXTENSIONS = set(["jpeg", "jpg", "jpe", "png", "gif", "webp"])
+    ALLOWED_DOC_EXTENSIONS = set(["jpeg", "jpg", "jpe", "png", "gif", "webp",
+        "pdf", "doc", "docx", "xls", "xlxs", "ppt", "pptx",
     ])
     # User settings
-    APPROVE_REGISTRATIONS = os.environ.get('APPROVE_REGISTRATIONS', False)
+    APPROVE_REGISTRATIONS = os.environ.get("APPROVE_REGISTRATIONS", False)
 
     SEED = 666
 
@@ -56,32 +58,32 @@ class BaseConfig(object):
 class DevelopmentConfig(BaseConfig):
     DEBUG = True
     SQLALCHEMY_ECHO = True
-    SQLALCHEMY_DATABASE_URI = os.environ.get('SQLALCHEMY_DATABASE_URI')
+    SQLALCHEMY_DATABASE_URI = os.environ.get("SQLALCHEMY_DATABASE_URI")
     NUM_OFFICERS = 15000
-    SITEMAP_URL_SCHEME = 'http'
+    SITEMAP_URL_SCHEME = "http"
 
 
 class TestingConfig(BaseConfig):
     TESTING = True
-    SQLALCHEMY_DATABASE_URI = 'sqlite:///:memory:'
+    SQLALCHEMY_DATABASE_URI = "sqlite:///:memory:"
     WTF_CSRF_ENABLED = False
     NUM_OFFICERS = 120
     APPROVE_REGISTRATIONS = False
-    SITEMAP_URL_SCHEME = 'http'
+    SITEMAP_URL_SCHEME = "http"
 
 
 class ProductionConfig(BaseConfig):
-    SQLALCHEMY_DATABASE_URI = os.environ.get('SQLALCHEMY_DATABASE_URI')
-    SITEMAP_URL_SCHEME = 'https'
+    SQLALCHEMY_DATABASE_URI = os.environ.get("SQLALCHEMY_DATABASE_URI")
+    SITEMAP_URL_SCHEME = "https"
 
     @classmethod
     def init_app(cls, app):  # pragma: no cover
-        BaseConfig.init_app(app)
+        config.init_app(app)
 
 
 config = {
-    'development': DevelopmentConfig,
-    'testing': TestingConfig,
-    'production': ProductionConfig,
+    "development": DevelopmentConfig,
+    "testing": TestingConfig,
+    "production": ProductionConfig,
 }
-config['default'] = config.get(os.environ.get('FLASK_ENV', ""), DevelopmentConfig)
+config["default"] = config.get(os.environ.get("FLASK_ENV", ""), DevelopmentConfig)
