@@ -3,6 +3,6 @@ import os
 
 os.environ['wsgi.url_scheme'] = 'https'
 
-main = Blueprint('main', __name__)  # noqa
+main = Blueprint("main", __name__)
 
-from . import views  # noqa
+from . import views  # noqa: E402,F401
