@@ -2,6 +2,8 @@ from typing import TYPE_CHECKING, Any, Dict, Optional, Sequence, Tuple, Union
 
 import dateutil.parser
 
+from OpenOversight.app.utils.general import get_or_create, str_is_true
+
 from .main import choices
 from .models import (
     Assignment,
@@ -13,11 +15,11 @@ from .models import (
     Salary,
     db,
 )
-from .utils import get_or_create, str_is_true
 from .validators import state_validator, url_validator
 
+
 if TYPE_CHECKING:
-    import datetime  # noqa
+    import datetime
 
 
 def validate_choice(
@@ -238,7 +240,12 @@ def get_or_create_license_plate_from_dict(
     number = data["number"]
     state = parse_str(data.get("state"), None)
     state_validator(state)
-    return get_or_create(db.session, LicensePlate, number=number, state=state,)
+    return get_or_create(
+        db.session,
+        LicensePlate,
+        number=number,
+        state=state,
+    )
 
 
 def get_or_create_location_from_dict(
