@@ -5,46 +5,65 @@ Revises: af933dc1ef93
 Create Date: 2018-05-03 15:00:36.849627
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 from sqlalchemy.dialects import postgresql
 
+
 # revision identifiers, used by Alembic.
-revision = '0acbb0f0b1ef'
-down_revision = '0ed957db0058'
+revision = "0acbb0f0b1ef"
+down_revision = "0ed957db0058"
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('assignments', 'resign_date',
-                    existing_type=postgresql.TIMESTAMP(),
-                    type_=sa.Date(),
-                    existing_nullable=True)
-    op.alter_column('assignments', 'star_date',
-                    existing_type=postgresql.TIMESTAMP(),
-                    type_=sa.Date(),
-                    existing_nullable=True)
-    op.alter_column('officers', 'employment_date',
-                    existing_type=postgresql.TIMESTAMP(),
-                    type_=sa.Date(),
-                    existing_nullable=True)
+    op.alter_column(
+        "assignments",
+        "resign_date",
+        existing_type=postgresql.TIMESTAMP(),
+        type_=sa.Date(),
+        existing_nullable=True,
+    )
+    op.alter_column(
+        "assignments",
+        "star_date",
+        existing_type=postgresql.TIMESTAMP(),
+        type_=sa.Date(),
+        existing_nullable=True,
+    )
+    op.alter_column(
+        "officers",
+        "employment_date",
+        existing_type=postgresql.TIMESTAMP(),
+        type_=sa.Date(),
+        existing_nullable=True,
+    )
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('officers', 'employment_date',
-                    existing_type=sa.Date(),
-                    type_=postgresql.TIMESTAMP(),
-                    existing_nullable=True)
-    op.alter_column('assignments', 'star_date',
-                    existing_type=sa.Date(),
-                    type_=postgresql.TIMESTAMP(),
-                    existing_nullable=True)
-    op.alter_column('assignments', 'resign_date',
-                    existing_type=sa.Date(),
-                    type_=postgresql.TIMESTAMP(),
-                    existing_nullable=True)
+    op.alter_column(
+        "officers",
+        "employment_date",
+        existing_type=sa.Date(),
+        type_=postgresql.TIMESTAMP(),
+        existing_nullable=True,
+    )
+    op.alter_column(
+        "assignments",
+        "star_date",
+        existing_type=sa.Date(),
+        type_=postgresql.TIMESTAMP(),
+        existing_nullable=True,
+    )
+    op.alter_column(
+        "assignments",
+        "resign_date",
+        existing_type=sa.Date(),
+        type_=postgresql.TIMESTAMP(),
+        existing_nullable=True,
+    )
     # ### end Alembic commands ###
