@@ -5,36 +5,37 @@ Revises: 7bb53dee8ac9
 Create Date: 2018-08-11 20:31:02.265231
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 
 
 # revision identifiers, used by Alembic.
-revision = '0ed957db0058'
-down_revision = '2c27bfebe66e'
+revision = "0ed957db0058"
+down_revision = "2c27bfebe66e"
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('descriptions',
-                    sa.Column('id', sa.Integer(), nullable=False),
-                    sa.Column('text_contents', sa.Text(), nullable=True),
-                    sa.Column('creator_id', sa.Integer(), nullable=True),
-                    sa.Column('officer_id', sa.Integer(), nullable=True),
-                    sa.Column('date_created', sa.DateTime(), nullable=True),
-                    sa.Column('date_updated', sa.DateTime(), nullable=True),
-                    sa.ForeignKeyConstraint(['creator_id'], ['users.id'], ondelete='SET NULL'),
-                    sa.ForeignKeyConstraint(['officer_id'], ['officers.id'], ondelete='CASCADE'),
-                    sa.PrimaryKeyConstraint('id')
-                    )
-    op.alter_column('notes', 'note', new_column_name='text_contents')
+    op.create_table(
+        "descriptions",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("text_contents", sa.Text(), nullable=True),
+        sa.Column("creator_id", sa.Integer(), nullable=True),
+        sa.Column("officer_id", sa.Integer(), nullable=True),
+        sa.Column("date_created", sa.DateTime(), nullable=True),
+        sa.Column("date_updated", sa.DateTime(), nullable=True),
+        sa.ForeignKeyConstraint(["creator_id"], ["users.id"], ondelete="SET NULL"),
+        sa.ForeignKeyConstraint(["officer_id"], ["officers.id"], ondelete="CASCADE"),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.alter_column("notes", "note", new_column_name="text_contents")
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('notes', 'text_contents', new_column_name='note')
-    op.drop_table('descriptions')
+    op.alter_column("notes", "text_contents", new_column_name="note")
+    op.drop_table("descriptions")
     # ### end Alembic commands ###
