@@ -5,13 +5,13 @@ Revises: bd0398fe4aab
 Create Date: 2018-06-06 19:34:16.439093
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 
 
 # revision identifiers, used by Alembic.
-revision = '2040f0c804b0'
-down_revision = 'bd0398fe4aab'
+revision = "2040f0c804b0"
+down_revision = "bd0398fe4aab"
 branch_labels = None
 depends_on = None
 
@@ -19,21 +19,27 @@ depends_on = None
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
     op.create_table(
-        'notes',
-        sa.Column('id', sa.Integer(), nullable=False),
-        sa.Column('note', sa.Text(), nullable=True),
-        sa.Column('user_id', sa.Integer(), nullable=True),
-        sa.Column('officer_id', sa.Integer(), nullable=True),
-        sa.Column('date_created', sa.DateTime(), nullable=True),
-        sa.Column('date_updated', sa.DateTime(), nullable=True),
-        sa.ForeignKeyConstraint(['officer_id'], ['officers.id'], ),
-        sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-        sa.PrimaryKeyConstraint('id')
+        "notes",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("note", sa.Text(), nullable=True),
+        sa.Column("user_id", sa.Integer(), nullable=True),
+        sa.Column("officer_id", sa.Integer(), nullable=True),
+        sa.Column("date_created", sa.DateTime(), nullable=True),
+        sa.Column("date_updated", sa.DateTime(), nullable=True),
+        sa.ForeignKeyConstraint(
+            ["officer_id"],
+            ["officers.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["user_id"],
+            ["users.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
     )
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('notes')
+    op.drop_table("notes")
     # ### end Alembic commands ###
