@@ -5,26 +5,34 @@ Revises: 7bb53dee8ac9
 Create Date: 2018-08-18 13:36:00.987327
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 
 
 # revision identifiers, used by Alembic.
-revision = '2c27bfebe66e'
-down_revision = '7bb53dee8ac9'
+revision = "2c27bfebe66e"
+down_revision = "7bb53dee8ac9"
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('officers', sa.Column('unique_internal_identifier', sa.String(length=50), nullable=True))
-    op.create_index(op.f('ix_officers_unique_internal_identifier'), 'officers', ['unique_internal_identifier'], unique=True)
+    op.add_column(
+        "officers",
+        sa.Column("unique_internal_identifier", sa.String(length=50), nullable=True),
+    )
+    op.create_index(
+        op.f("ix_officers_unique_internal_identifier"),
+        "officers",
+        ["unique_internal_identifier"],
+        unique=True,
+    )
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_officers_unique_internal_identifier'), table_name='officers')
-    op.drop_column('officers', 'unique_internal_identifier')
+    op.drop_index(op.f("ix_officers_unique_internal_identifier"), table_name="officers")
+    op.drop_column("officers", "unique_internal_identifier")
     # ### end Alembic commands ###
