@@ -5,30 +5,36 @@ Revises: 114919b27a9f
 Create Date: 2017-12-23 18:03:07.304611
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 
 
 # revision identifiers, used by Alembic.
-revision = '42233d18ac7b'
-down_revision = '114919b27a9f'
+revision = "42233d18ac7b"
+down_revision = "114919b27a9f"
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('assignments', 'star_no',
-                    existing_type=sa.INTEGER(),
-                    type_=sa.String(length=120),
-                    existing_nullable=True)
+    op.alter_column(
+        "assignments",
+        "star_no",
+        existing_type=sa.INTEGER(),
+        type_=sa.String(length=120),
+        existing_nullable=True,
+    )
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('assignments', 'star_no',
-                    existing_type=sa.String(length=120),
-                    type_=sa.INTEGER(),
-                    existing_nullable=True)
+    op.alter_column(
+        "assignments",
+        "star_no",
+        existing_type=sa.String(length=120),
+        type_=sa.INTEGER(),
+        existing_nullable=True,
+    )
     # ### end Alembic commands ###
