@@ -5,19 +5,22 @@ Revises: bd0398fe4aab
 Create Date: 2018-06-04 19:04:23.524079
 
 """
-from flask import current_app
 import os
 import sys
+
+from flask import current_app
+
+
 # Add our Flask app to the search paths for modules
 sys.path.insert(0, os.path.dirname(current_app.root_path))
 
 from app.models import Face, db  # noqa: E402
-from app.utils import crop_image  # noqa: E402
+from app.utils.cloud import crop_image  # noqa: E402
 
 
 # revision identifiers, used by Alembic.
-revision = '59e9993c169c'
-down_revision = '4a490771dda1'
+revision = "59e9993c169c"
+down_revision = "4a490771dda1"
 branch_labels = None
 depends_on = None
 
@@ -25,8 +28,7 @@ depends_on = None
 def upgrade():
     try:
         for face in Face.query.all():
-            if face.face_position_x \
-               and face.image.filepath.split('/')[0] != 'static':
+            if face.face_position_x and face.image.filepath.split("/")[0] != "static":
                 left = face.face_position_x
                 upper = face.face_position_y
                 right = left + face.face_width
@@ -42,7 +44,8 @@ def upgrade():
                     face_position_y=face.face_position_y,
                     face_height=face.face_height,
                     face_width=face.face_width,
-                    user_id=face.user_id)
+                    user_id=face.user_id,
+                )
 
                 db.session.add(cropped_image)
                 db.session.add(new_face)
