@@ -5,25 +5,25 @@ Revises: 9e2827dae28c
 Create Date: 2019-01-24 15:54:08.123125
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 
 
 # revision identifiers, used by Alembic.
-revision = '5c5b80cab45e'
-down_revision = 'e2c2efde8b55'
+revision = "5c5b80cab45e"
+down_revision = "e2c2efde8b55"
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('approved', sa.Boolean(), default=False))
+    op.add_column("users", sa.Column("approved", sa.Boolean(), default=False))
     op.execute("UPDATE users SET approved=True")
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('users', 'approved')
+    op.drop_column("users", "approved")
     # ### end Alembic commands ###
