@@ -5,38 +5,49 @@ Revises: 8ce3de7679c2
 Create Date: 2019-05-04 05:28:06.869101
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 from sqlalchemy.dialects import postgresql
 
+
 # revision identifiers, used by Alembic.
-revision = '6045f42587ec'
-down_revision = '8ce3de7679c2'
+revision = "6045f42587ec"
+down_revision = "8ce3de7679c2"
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('incidents', sa.Column('time', sa.Time(), nullable=True))
-    op.execute('UPDATE incidents SET time = date::time')
-    op.alter_column('incidents', 'date',
-                    existing_type=postgresql.TIMESTAMP(),
-                    type_=sa.Date(),
-                    existing_nullable=True)
-    op.create_index(op.f('ix_incidents_time'), 'incidents', ['time'], unique=False)
-    op.execute('UPDATE incidents SET "time" = NULL WHERE "time" = \'01:02:03.045678\'::time')
+    op.add_column("incidents", sa.Column("time", sa.Time(), nullable=True))
+    op.execute("UPDATE incidents SET time = date::time")
+    op.alter_column(
+        "incidents",
+        "date",
+        existing_type=postgresql.TIMESTAMP(),
+        type_=sa.Date(),
+        existing_nullable=True,
+    )
+    op.create_index(op.f("ix_incidents_time"), "incidents", ["time"], unique=False)
+    op.execute(
+        'UPDATE incidents SET "time" = NULL WHERE "time" = \'01:02:03.045678\'::time'
+    )
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.execute('UPDATE incidents SET "time" = \'01:02:03.045678\'::time WHERE time IS NULL')
-    op.drop_index(op.f('ix_incidents_time'), table_name='incidents')
-    op.alter_column('incidents', 'date',
-                    existing_type=sa.Date(),
-                    type_=postgresql.TIMESTAMP(),
-                    existing_nullable=True)
-    op.execute('UPDATE incidents SET date = date + time')
-    op.drop_column('incidents', 'time')
+    op.execute(
+        "UPDATE incidents SET \"time\" = '01:02:03.045678'::time WHERE time IS NULL"
+    )
+    op.drop_index(op.f("ix_incidents_time"), table_name="incidents")
+    op.alter_column(
+        "incidents",
+        "date",
+        existing_type=sa.Date(),
+        type_=postgresql.TIMESTAMP(),
+        existing_nullable=True,
+    )
+    op.execute("UPDATE incidents SET date = date + time")
+    op.drop_column("incidents", "time")
     # ### end Alembic commands ###
