@@ -5,13 +5,13 @@ Revises: d86feb8fa5d1
 Create Date: 2018-05-04 21:05:42.060165
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 
 
 # revision identifiers, used by Alembic.
-revision = '6065d7cdcbf8'
-down_revision = 'd86feb8fa5d1'
+revision = "6065d7cdcbf8"
+down_revision = "d86feb8fa5d1"
 branch_labels = None
 depends_on = None
 
@@ -19,25 +19,49 @@ depends_on = None
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
     op.create_table(
-        'incident_officers',
-        sa.Column('incident_id', sa.Integer(), nullable=False),
-        sa.Column('officers_id', sa.Integer(), nullable=False),
-        sa.ForeignKeyConstraint(['incident_id'], ['incidents.id'], ),
-        sa.ForeignKeyConstraint(['officers_id'], ['officers.id'], ),
-        sa.PrimaryKeyConstraint('incident_id', 'officers_id')
+        "incident_officers",
+        sa.Column("incident_id", sa.Integer(), nullable=False),
+        sa.Column("officers_id", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["incident_id"],
+            ["incidents.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["officers_id"],
+            ["officers.id"],
+        ),
+        sa.PrimaryKeyConstraint("incident_id", "officers_id"),
     )
-    op.add_column(u'incident_license_plates', sa.Column('license_plate_id', sa.Integer(), nullable=False))
-    op.drop_constraint(u'incident_license_plates_link_id_fkey', 'incident_license_plates', type_='foreignkey')
-    op.create_foreign_key(None, 'incident_license_plates', 'license_plates', ['license_plate_id'], ['id'])
-    op.drop_column(u'incident_license_plates', 'link_id')
+    op.add_column(
+        "incident_license_plates",
+        sa.Column("license_plate_id", sa.Integer(), nullable=False),
+    )
+    op.drop_constraint(
+        "incident_license_plates_link_id_fkey",
+        "incident_license_plates",
+        type_="foreignkey",
+    )
+    op.create_foreign_key(
+        None, "incident_license_plates", "license_plates", ["license_plate_id"], ["id"]
+    )
+    op.drop_column("incident_license_plates", "link_id")
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column(u'incident_license_plates', sa.Column('link_id', sa.INTEGER(), autoincrement=False, nullable=False))
-    op.drop_constraint(None, 'incident_license_plates', type_='foreignkey')
-    op.create_foreign_key(u'incident_license_plates_link_id_fkey', 'incident_license_plates', 'license_plates', ['link_id'], ['id'])
-    op.drop_column(u'incident_license_plates', 'license_plate_id')
-    op.drop_table('incident_officers')
+    op.add_column(
+        "incident_license_plates",
+        sa.Column("link_id", sa.INTEGER(), autoincrement=False, nullable=False),
+    )
+    op.drop_constraint(None, "incident_license_plates", type_="foreignkey")
+    op.create_foreign_key(
+        "incident_license_plates_link_id_fkey",
+        "incident_license_plates",
+        "license_plates",
+        ["link_id"],
+        ["id"],
+    )
+    op.drop_column("incident_license_plates", "license_plate_id")
+    op.drop_table("incident_officers")
     # ### end Alembic commands ###
