@@ -5,24 +5,27 @@ Revises: 562bd5f1bc1f
 Create Date: 2020-07-31 14:03:59.871182
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 
 
 # revision identifiers, used by Alembic.
-revision = '79a14685454f'
-down_revision = '562bd5f1bc1f'
+revision = "79a14685454f"
+down_revision = "562bd5f1bc1f"
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('faces', sa.Column('featured', sa.Boolean(), server_default='false', nullable=False))
+    op.add_column(
+        "faces",
+        sa.Column("featured", sa.Boolean(), server_default="false", nullable=False),
+    )
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('faces', 'featured')
+    op.drop_column("faces", "featured")
     # ### end Alembic commands ###
