@@ -5,26 +5,26 @@ Revises: 59e9993c169c
 Create Date: 2018-07-26 18:36:10.061968
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 
 
 # revision identifiers, used by Alembic.
-revision = '7bb53dee8ac9'
-down_revision = '59e9993c169c'
+revision = "7bb53dee8ac9"
+down_revision = "59e9993c169c"
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('officers', sa.Column('suffix', sa.String(length=120), nullable=True))
-    op.create_index(op.f('ix_officers_suffix'), 'officers', ['suffix'], unique=False)
+    op.add_column("officers", sa.Column("suffix", sa.String(length=120), nullable=True))
+    op.create_index(op.f("ix_officers_suffix"), "officers", ["suffix"], unique=False)
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_officers_suffix'), table_name='officers')
-    op.drop_column('officers', 'suffix')
+    op.drop_index(op.f("ix_officers_suffix"), table_name="officers")
+    op.drop_column("officers", "suffix")
     # ### end Alembic commands ###
