@@ -5,24 +5,29 @@ Revises: 3015d1dd9eb4
 Create Date: 2019-04-17 18:26:35.733783
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 
 
 # revision identifiers, used by Alembic.
-revision = '8ce3de7679c2'
-down_revision = '3015d1dd9eb4'
+revision = "8ce3de7679c2"
+down_revision = "3015d1dd9eb4"
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('departments', sa.Column('unique_internal_identifier_label', sa.String(length=100), nullable=True))
+    op.add_column(
+        "departments",
+        sa.Column(
+            "unique_internal_identifier_label", sa.String(length=100), nullable=True
+        ),
+    )
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('departments', 'unique_internal_identifier_label')
+    op.drop_column("departments", "unique_internal_identifier_label")
     # ### end Alembic commands ###
