@@ -9,25 +9,33 @@ from alembic import op
 
 
 # revision identifiers, used by Alembic.
-revision = '8ce7926aa132'
-down_revision = 'cfc5f3fd5efe'
+revision = "8ce7926aa132"
+down_revision = "cfc5f3fd5efe"
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(u'notes_officer_id_fkey', 'notes', type_='foreignkey')
-    op.drop_constraint(u'notes_creator_id_fkey', 'notes', type_='foreignkey')
-    op.create_foreign_key(None, 'notes', 'officers', ['officer_id'], ['id'], ondelete='CASCADE')
-    op.create_foreign_key(None, 'notes', 'users', ['creator_id'], ['id'], ondelete='SET NULL')
+    op.drop_constraint("notes_officer_id_fkey", "notes", type_="foreignkey")
+    op.drop_constraint("notes_creator_id_fkey", "notes", type_="foreignkey")
+    op.create_foreign_key(
+        None, "notes", "officers", ["officer_id"], ["id"], ondelete="CASCADE"
+    )
+    op.create_foreign_key(
+        None, "notes", "users", ["creator_id"], ["id"], ondelete="SET NULL"
+    )
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'notes', type_='foreignkey')
-    op.drop_constraint(None, 'notes', type_='foreignkey')
-    op.create_foreign_key(u'notes_creator_id_fkey', 'notes', 'users', ['creator_id'], ['id'])
-    op.create_foreign_key(u'notes_officer_id_fkey', 'notes', 'officers', ['officer_id'], ['id'])
+    op.drop_constraint(None, "notes", type_="foreignkey")
+    op.drop_constraint(None, "notes", type_="foreignkey")
+    op.create_foreign_key(
+        "notes_creator_id_fkey", "notes", "users", ["creator_id"], ["id"]
+    )
+    op.create_foreign_key(
+        "notes_officer_id_fkey", "notes", "officers", ["officer_id"], ["id"]
+    )
     # ### end Alembic commands ###
