@@ -5,26 +5,26 @@ Revises: 42233d18ac7b
 Create Date: 2018-04-12 15:41:33.490603
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 
 
 # revision identifiers, used by Alembic.
-revision = 'af933dc1ef93'
-down_revision = '42233d18ac7b'
+revision = "af933dc1ef93"
+down_revision = "42233d18ac7b"
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('dept_pref', sa.Integer(), nullable=True))
-    op.create_foreign_key(None, 'users', 'departments', ['dept_pref'], ['id'])
+    op.add_column("users", sa.Column("dept_pref", sa.Integer(), nullable=True))
+    op.create_foreign_key(None, "users", "departments", ["dept_pref"], ["id"])
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'users', type_='foreignkey')
-    op.drop_column('users', 'dept_pref')
+    op.drop_constraint(None, "users", type_="foreignkey")
+    op.drop_column("users", "dept_pref")
     # ### end Alembic commands ###
