@@ -5,32 +5,40 @@ Revises: f4a41e328a06
 Create Date: 2018-05-15 19:56:14.199692
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 
 
 # revision identifiers, used by Alembic.
-revision = 'bd0398fe4aab'
-down_revision = 'f4a41e328a06'
+revision = "bd0398fe4aab"
+down_revision = "f4a41e328a06"
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('incidents', 'user_id', new_column_name='creator_id', existing_type=sa.Integer)
-    op.add_column('incidents', sa.Column('last_updated_id', sa.Integer(), nullable=True))
-    op.drop_constraint(u'incidents_user_id_fkey', 'incidents', type_='foreignkey')
-    op.create_foreign_key(None, 'incidents', 'users', ['creator_id'], ['id'])
-    op.create_foreign_key(None, 'incidents', 'users', ['last_updated_id'], ['id'])
+    op.alter_column(
+        "incidents", "user_id", new_column_name="creator_id", existing_type=sa.Integer
+    )
+    op.add_column(
+        "incidents", sa.Column("last_updated_id", sa.Integer(), nullable=True)
+    )
+    op.drop_constraint("incidents_user_id_fkey", "incidents", type_="foreignkey")
+    op.create_foreign_key(None, "incidents", "users", ["creator_id"], ["id"])
+    op.create_foreign_key(None, "incidents", "users", ["last_updated_id"], ["id"])
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('incidents', 'creator_id', new_column_name='user_id', existing_type=sa.Integer)
-    op.drop_constraint(None, 'incidents', type_='foreignkey')
-    op.drop_constraint(None, 'incidents', type_='foreignkey')
-    op.create_foreign_key(u'incidents_user_id_fkey', 'incidents', 'users', ['user_id'], ['id'])
-    op.drop_column('incidents', 'last_updated_id')
+    op.alter_column(
+        "incidents", "creator_id", new_column_name="user_id", existing_type=sa.Integer
+    )
+    op.drop_constraint(None, "incidents", type_="foreignkey")
+    op.drop_constraint(None, "incidents", type_="foreignkey")
+    op.create_foreign_key(
+        "incidents_user_id_fkey", "incidents", "users", ["user_id"], ["id"]
+    )
+    op.drop_column("incidents", "last_updated_id")
     # ### end Alembic commands ###
