@@ -9,8 +9,8 @@ from alembic import op
 
 
 # revision identifiers, used by Alembic.
-revision = 'c1fc26073f85'
-down_revision = '770ed51b4e16'
+revision = "c1fc26073f85"
+down_revision = "770ed51b4e16"
 branch_labels = None
 depends_on = None
 
