@@ -5,26 +5,26 @@ Revises: 6065d7cdcbf8
 Create Date: 2018-05-10 20:02:19.006081
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 
 
 # revision identifiers, used by Alembic.
-revision = 'ca95c047bf42'
-down_revision = '6065d7cdcbf8'
+revision = "ca95c047bf42"
+down_revision = "6065d7cdcbf8"
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('incidents', sa.Column('department_id', sa.Integer(), nullable=True))
-    op.create_foreign_key(None, 'incidents', 'departments', ['department_id'], ['id'])
+    op.add_column("incidents", sa.Column("department_id", sa.Integer(), nullable=True))
+    op.create_foreign_key(None, "incidents", "departments", ["department_id"], ["id"])
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'incidents', type_='foreignkey')
-    op.drop_column('incidents', 'department_id')
+    op.drop_constraint(None, "incidents", type_="foreignkey")
+    op.drop_column("incidents", "department_id")
     # ### end Alembic commands ###
