@@ -5,13 +5,13 @@ Revises: 86eb228e4bc0
 Create Date: 2020-07-13 02:45:07.533549
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 
 
 # revision identifiers, used by Alembic.
-revision = 'cd39b33b5360'
-down_revision = '86eb228e4bc0'
+revision = "cd39b33b5360"
+down_revision = "86eb228e4bc0"
 branch_labels = None
 depends_on = None
 
@@ -22,7 +22,7 @@ def get_update_statement(normalized, options):
     SET gender = '{normalized}'
     WHERE LOWER(gender) in ({options});
     """
-    options = ', '.join(["'" + o + "'" for o in options])
+    options = ", ".join(["'" + o + "'" for o in options])
     return template.format(normalized=normalized, options=options)
 
 
@@ -30,12 +30,12 @@ def upgrade():
     conn = op.get_bind()
 
     genders = {
-        "M": ('male', 'm', 'man'),
-        "F": ('female', 'f', 'woman'),
-        "Other": ('nonbinary', 'other'),
+        "M": ("male", "m", "man"),
+        "F": ("female", "f", "woman"),
+        "Other": ("nonbinary", "other"),
     }
 
-    update_statement = ''
+    update_statement = ""
 
     for normalized, options in genders.items():
         update_statement += get_update_statement(normalized, options)
@@ -50,25 +50,29 @@ WHERE gender not in ('M', 'F', 'Other');
     conn.execute(null_query)
 
     # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('officers', 'gender',
-                    existing_type=sa.VARCHAR(length=120),
-                    type_=sa.VARCHAR(length=5),
-                    existing_nullable=True)
+    op.alter_column(
+        "officers",
+        "gender",
+        existing_type=sa.VARCHAR(length=120),
+        type_=sa.VARCHAR(length=5),
+        existing_nullable=True,
+    )
     # ### end Alembic commands ###
 
     op.create_check_constraint(
-        'gender_options',
-        'officers',
-        "gender in ('M', 'F', 'Other')"
+        "gender_options", "officers", "gender in ('M', 'F', 'Other')"
     )
 
 
 def downgrade():
-    op.drop_constraint('gender_options', 'officers', type_='check')
+    op.drop_constraint("gender_options", "officers", type_="check")
 
     # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('officers', 'gender',
-                    existing_type=sa.VARCHAR(length=5),
-                    type_=sa.VARCHAR(length=120),
-                    existing_nullable=True)
+    op.alter_column(
+        "officers",
+        "gender",
+        existing_type=sa.VARCHAR(length=5),
+        type_=sa.VARCHAR(length=120),
+        existing_nullable=True,
+    )
     # ### end Alembic commands ###
