@@ -5,13 +5,13 @@ Revises: e14a1aa4b58f
 Create Date: 2018-05-04 21:03:12.925484
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 
 
 # revision identifiers, used by Alembic.
-revision = 'd86feb8fa5d1'
-down_revision = 'e14a1aa4b58f'
+revision = "d86feb8fa5d1"
+down_revision = "e14a1aa4b58f"
 branch_labels = None
 depends_on = None
 
@@ -19,101 +19,138 @@ depends_on = None
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
     op.create_table(
-        'license_plates',
-        sa.Column('id', sa.Integer(), nullable=False),
-        sa.Column('number', sa.String(length=8), nullable=False),
-        sa.Column('state', sa.String(length=2), nullable=True),
-        sa.PrimaryKeyConstraint('id')
+        "license_plates",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("number", sa.String(length=8), nullable=False),
+        sa.Column("state", sa.String(length=2), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
     )
-    op.create_index(op.f('ix_license_plates_number'), 'license_plates', ['number'], unique=False)
-    op.create_index(op.f('ix_license_plates_state'), 'license_plates', ['state'], unique=False)
-    op.create_table(
-        'links',
-        sa.Column('id', sa.Integer(), nullable=False),
-        sa.Column('url', sa.String(length=255), nullable=False),
-        sa.Column('link_type', sa.String(length=100), nullable=True),
-        sa.PrimaryKeyConstraint('id')
+    op.create_index(
+        op.f("ix_license_plates_number"), "license_plates", ["number"], unique=False
     )
-    op.create_index(op.f('ix_links_link_type'), 'links', ['link_type'], unique=False)
-    op.create_table(
-        'locations',
-        sa.Column('id', sa.Integer(), nullable=False),
-        sa.Column('street_name', sa.String(length=100), nullable=True),
-        sa.Column('cross_street1', sa.String(length=100), nullable=True),
-        sa.Column('cross_street2', sa.String(length=100), nullable=True),
-        sa.Column('city', sa.String(length=100), nullable=True),
-        sa.Column('state', sa.String(length=2), nullable=True),
-        sa.Column('zip_code', sa.String(length=5), nullable=True),
-        sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_locations_city'), 'locations', ['city'], unique=False)
-    op.create_index(op.f('ix_locations_state'), 'locations', ['state'], unique=False)
-    op.create_index(op.f('ix_locations_street_name'), 'locations', ['street_name'], unique=False)
-    op.create_index(op.f('ix_locations_zip_code'), 'locations', ['zip_code'], unique=False)
-    op.create_table(
-        'incidents',
-        sa.Column('id', sa.Integer(), nullable=False),
-        sa.Column('date', sa.DateTime(), nullable=True),
-        sa.Column('report_number', sa.String(length=50), nullable=True),
-        sa.Column('description', sa.Text(), nullable=True),
-        sa.Column('address_id', sa.Integer(), nullable=True),
-        sa.ForeignKeyConstraint(['address_id'], ['locations.id'], ),
-        sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_incidents_date'), 'incidents', ['date'], unique=False)
-    op.create_index(op.f('ix_incidents_report_number'), 'incidents', ['report_number'], unique=False)
-    op.create_table(
-        'incident_license_plates',
-        sa.Column('incident_id', sa.Integer(), nullable=False),
-        sa.Column('link_id', sa.Integer(), nullable=False),
-        sa.ForeignKeyConstraint(['incident_id'], ['incidents.id'], ),
-        sa.ForeignKeyConstraint(['link_id'], ['license_plates.id'], ),
-        sa.PrimaryKeyConstraint('incident_id', 'link_id')
+    op.create_index(
+        op.f("ix_license_plates_state"), "license_plates", ["state"], unique=False
     )
     op.create_table(
-        'incident_links',
-        sa.Column('incident_id', sa.Integer(), nullable=False),
-        sa.Column('link_id', sa.Integer(), nullable=False),
-        sa.ForeignKeyConstraint(['incident_id'], ['incidents.id'], ),
-        sa.ForeignKeyConstraint(['link_id'], ['links.id'], ),
-        sa.PrimaryKeyConstraint('incident_id', 'link_id')
+        "links",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("url", sa.String(length=255), nullable=False),
+        sa.Column("link_type", sa.String(length=100), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_links_link_type"), "links", ["link_type"], unique=False)
+    op.create_table(
+        "locations",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("street_name", sa.String(length=100), nullable=True),
+        sa.Column("cross_street1", sa.String(length=100), nullable=True),
+        sa.Column("cross_street2", sa.String(length=100), nullable=True),
+        sa.Column("city", sa.String(length=100), nullable=True),
+        sa.Column("state", sa.String(length=2), nullable=True),
+        sa.Column("zip_code", sa.String(length=5), nullable=True),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_locations_city"), "locations", ["city"], unique=False)
+    op.create_index(op.f("ix_locations_state"), "locations", ["state"], unique=False)
+    op.create_index(
+        op.f("ix_locations_street_name"), "locations", ["street_name"], unique=False
+    )
+    op.create_index(
+        op.f("ix_locations_zip_code"), "locations", ["zip_code"], unique=False
     )
     op.create_table(
-        'officer_incidents',
-        sa.Column('officer_id', sa.Integer(), nullable=False),
-        sa.Column('incident_id', sa.Integer(), nullable=False),
-        sa.ForeignKeyConstraint(['incident_id'], ['incidents.id'], ),
-        sa.ForeignKeyConstraint(['officer_id'], ['officers.id'], ),
-        sa.PrimaryKeyConstraint('officer_id', 'incident_id')
+        "incidents",
+        sa.Column("id", sa.Integer(), nullable=False),
+        sa.Column("date", sa.DateTime(), nullable=True),
+        sa.Column("report_number", sa.String(length=50), nullable=True),
+        sa.Column("description", sa.Text(), nullable=True),
+        sa.Column("address_id", sa.Integer(), nullable=True),
+        sa.ForeignKeyConstraint(
+            ["address_id"],
+            ["locations.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.create_index(op.f("ix_incidents_date"), "incidents", ["date"], unique=False)
+    op.create_index(
+        op.f("ix_incidents_report_number"), "incidents", ["report_number"], unique=False
     )
     op.create_table(
-        'officer_links',
-        sa.Column('officer_id', sa.Integer(), nullable=False),
-        sa.Column('link_id', sa.Integer(), nullable=False),
-        sa.ForeignKeyConstraint(['link_id'], ['links.id'], ),
-        sa.ForeignKeyConstraint(['officer_id'], ['officers.id'], ),
-        sa.PrimaryKeyConstraint('officer_id', 'link_id')
+        "incident_license_plates",
+        sa.Column("incident_id", sa.Integer(), nullable=False),
+        sa.Column("link_id", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["incident_id"],
+            ["incidents.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["link_id"],
+            ["license_plates.id"],
+        ),
+        sa.PrimaryKeyConstraint("incident_id", "link_id"),
+    )
+    op.create_table(
+        "incident_links",
+        sa.Column("incident_id", sa.Integer(), nullable=False),
+        sa.Column("link_id", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["incident_id"],
+            ["incidents.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["link_id"],
+            ["links.id"],
+        ),
+        sa.PrimaryKeyConstraint("incident_id", "link_id"),
+    )
+    op.create_table(
+        "officer_incidents",
+        sa.Column("officer_id", sa.Integer(), nullable=False),
+        sa.Column("incident_id", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["incident_id"],
+            ["incidents.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["officer_id"],
+            ["officers.id"],
+        ),
+        sa.PrimaryKeyConstraint("officer_id", "incident_id"),
+    )
+    op.create_table(
+        "officer_links",
+        sa.Column("officer_id", sa.Integer(), nullable=False),
+        sa.Column("link_id", sa.Integer(), nullable=False),
+        sa.ForeignKeyConstraint(
+            ["link_id"],
+            ["links.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["officer_id"],
+            ["officers.id"],
+        ),
+        sa.PrimaryKeyConstraint("officer_id", "link_id"),
     )
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('officer_links')
-    op.drop_table('officer_incidents')
-    op.drop_table('incident_links')
-    op.drop_table('incident_license_plates')
-    op.drop_index(op.f('ix_incidents_report_number'), table_name='incidents')
-    op.drop_index(op.f('ix_incidents_date'), table_name='incidents')
-    op.drop_table('incidents')
-    op.drop_index(op.f('ix_locations_zip_code'), table_name='locations')
-    op.drop_index(op.f('ix_locations_street_name'), table_name='locations')
-    op.drop_index(op.f('ix_locations_state'), table_name='locations')
-    op.drop_index(op.f('ix_locations_city'), table_name='locations')
-    op.drop_table('locations')
-    op.drop_index(op.f('ix_links_link_type'), table_name='links')
-    op.drop_table('links')
-    op.drop_index(op.f('ix_license_plates_state'), table_name='license_plates')
-    op.drop_index(op.f('ix_license_plates_number'), table_name='license_plates')
-    op.drop_table('license_plates')
+    op.drop_table("officer_links")
+    op.drop_table("officer_incidents")
+    op.drop_table("incident_links")
+    op.drop_table("incident_license_plates")
+    op.drop_index(op.f("ix_incidents_report_number"), table_name="incidents")
+    op.drop_index(op.f("ix_incidents_date"), table_name="incidents")
+    op.drop_table("incidents")
+    op.drop_index(op.f("ix_locations_zip_code"), table_name="locations")
+    op.drop_index(op.f("ix_locations_street_name"), table_name="locations")
+    op.drop_index(op.f("ix_locations_state"), table_name="locations")
+    op.drop_index(op.f("ix_locations_city"), table_name="locations")
+    op.drop_table("locations")
+    op.drop_index(op.f("ix_links_link_type"), table_name="links")
+    op.drop_table("links")
+    op.drop_index(op.f("ix_license_plates_state"), table_name="license_plates")
+    op.drop_index(op.f("ix_license_plates_number"), table_name="license_plates")
+    op.drop_table("license_plates")
     # ### end Alembic commands ###
