@@ -5,28 +5,30 @@ Revises: af933dc1ef93
 Create Date: 2018-04-30 15:48:51.968189
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 
 
 # revision identifiers, used by Alembic.
-revision = 'e14a1aa4b58f'
-down_revision = 'af933dc1ef93'
+revision = "e14a1aa4b58f"
+down_revision = "af933dc1ef93"
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('ac_department_id', sa.Integer(), nullable=True))
-    op.add_column('users', sa.Column('is_area_coordinator', sa.Boolean(), nullable=True))
-    op.create_foreign_key(None, 'users', 'departments', ['ac_department_id'], ['id'])
+    op.add_column("users", sa.Column("ac_department_id", sa.Integer(), nullable=True))
+    op.add_column(
+        "users", sa.Column("is_area_coordinator", sa.Boolean(), nullable=True)
+    )
+    op.create_foreign_key(None, "users", "departments", ["ac_department_id"], ["id"])
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'users', type_='foreignkey')
-    op.drop_column('users', 'is_area_coordinator')
-    op.drop_column('users', 'ac_department_id')
+    op.drop_constraint(None, "users", type_="foreignkey")
+    op.drop_column("users", "is_area_coordinator")
+    op.drop_column("users", "ac_department_id")
     # ### end Alembic commands ###
