@@ -9,23 +9,46 @@ from alembic import op
 
 
 # revision identifiers, used by Alembic.
-revision = 'e2c2efde8b55'
-down_revision = '9e2827dae28c'
+revision = "e2c2efde8b55"
+down_revision = "9e2827dae28c"
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('faces', 'fk_face_original_image_id', new_column_name='original_image_id')
-    op.drop_constraint('faces_fk_face_original_image_id_fkey', 'faces', type_='foreignkey')
-    op.create_foreign_key('fk_face_original_image_id', 'faces', 'raw_images', ['original_image_id'], ['id'], onupdate='CASCADE', ondelete='SET NULL', use_alter=True)
+    op.alter_column(
+        "faces", "fk_face_original_image_id", new_column_name="original_image_id"
+    )
+    op.drop_constraint(
+        "faces_fk_face_original_image_id_fkey", "faces", type_="foreignkey"
+    )
+    op.create_foreign_key(
+        "fk_face_original_image_id",
+        "faces",
+        "raw_images",
+        ["original_image_id"],
+        ["id"],
+        onupdate="CASCADE",
+        ondelete="SET NULL",
+        use_alter=True,
+    )
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint('fk_face_original_image_id', 'faces', type_='foreignkey')
-    op.create_foreign_key('faces_fk_face_original_image_id_fkey', 'faces', 'raw_images', ['fk_face_original_image_id'], ['id'], onupdate='CASCADE', ondelete='SET NULL')
-    op.alter_column('faces', 'original_image_id', new_column_name='fk_face_original_image_id')
+    op.drop_constraint("fk_face_original_image_id", "faces", type_="foreignkey")
+    op.create_foreign_key(
+        "faces_fk_face_original_image_id_fkey",
+        "faces",
+        "raw_images",
+        ["fk_face_original_image_id"],
+        ["id"],
+        onupdate="CASCADE",
+        ondelete="SET NULL",
+    )
+    op.alter_column(
+        "faces", "original_image_id", new_column_name="fk_face_original_image_id"
+    )
     # ### end Alembic commands ###
