@@ -1,41 +1,50 @@
-import pytest
 from datetime import datetime
-from flask import url_for, current_app
+from http import HTTPStatus
+
+import pytest
+from flask import current_app, url_for
+
+from OpenOversight.app.main.forms import EditTextForm, TextForm
+from OpenOversight.app.models import Description, Officer, User, db
+from OpenOversight.app.utils.constants import ENCODING_UTF_8
 from OpenOversight.tests.conftest import AC_DEPT
-from .route_helpers import login_user, login_admin, login_ac
+
+from .route_helpers import login_ac, login_admin, login_user
 
 
-from OpenOversight.app.main.forms import TextForm, EditTextForm
-from OpenOversight.app.models import db, Officer, Description, User
-
-
-@pytest.mark.parametrize("route", [
-    ('officer/1/description/1/edit'),
-    ('officer/1/description/new'),
-    ('officer/1/description/1/delete')
-])
+@pytest.mark.parametrize(
+    "route",
+    [
+        "officer/1/description/1/edit",
+        "officer/1/description/new",
+        "officer/1/description/1/delete",
+    ],
+)
 def test_route_login_required(route, client, mockdata):
     rv = client.get(route)
-    assert rv.status_code == 302
+    assert rv.status_code == HTTPStatus.FOUND
 
 
-@pytest.mark.parametrize("route", [
-    ('officer/1/description/1/edit'),
-    ('officer/1/description/new'),
-    ('officer/1/description/1/delete')
-])
+@pytest.mark.parametrize(
+    "route",
+    [
+        "officer/1/description/1/edit",
+        "officer/1/description/new",
+        "officer/1/description/1/delete",
+    ],
+)
 def test_route_admin_or_required(route, client, mockdata):
     with current_app.test_request_context():
         login_user(client)
         rv = client.get(route)
-        assert rv.status_code == 403
+        assert rv.status_code == HTTPStatus.FORBIDDEN
 
 
 def test_officer_descriptions_markdown(mockdata, client, session):
     with current_app.test_request_context():
         login_user(client)
-        rv = client.get(url_for('main.officer_profile', officer_id=1))
-        assert rv.status_code == 200
+        rv = client.get(url_for("main.officer_profile", officer_id=1))
+        assert rv.status_code == HTTPStatus.OK
         html = rv.data.decode()
         print(html)
         assert "<h3>A markdown description</h3>" in html
@@ -46,22 +55,20 @@ def test_admins_cannot_inject_unsafe_html(mockdata, client, session):
     with current_app.test_request_context():
         login_admin(client)
         officer = Officer.query.first()
-        text_contents = 'New description\n<script>alert();</script>'
-        admin = User.query.filter_by(email='jen@example.org').first()
+        text_contents = "New description\n<script>alert();</script>"
+        admin = User.query.filter_by(email="jen@example.org").first()
         form = TextForm(
-            text_contents=text_contents,
-            officer_id=officer.id,
-            creator_id=admin.id
+            text_contents=text_contents, officer_id=officer.id, creator_id=admin.id
         )
 
         rv = client.post(
-            url_for('main.description_api', officer_id=officer.id),
+            url_for("main.description_api", officer_id=officer.id),
             data=form.data,
-            follow_redirects=True
+            follow_redirects=True,
         )
 
-        assert rv.status_code == 200
-        assert 'created' in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert "created" in rv.data.decode(ENCODING_UTF_8)
         assert "<script>" not in rv.data.decode()
         assert "&lt;script&gt;" in rv.data.decode()
 
@@ -70,24 +77,24 @@ def test_admins_can_create_descriptions(mockdata, client, session):
     with current_app.test_request_context():
         login_admin(client)
         officer = Officer.query.first()
-        text_contents = 'I can haz descriptionz'
-        admin = User.query.filter_by(email='jen@example.org').first()
+        text_contents = "I can haz descriptionz"
+        admin = User.query.filter_by(email="jen@example.org").first()
         form = TextForm(
-            text_contents=text_contents,
-            officer_id=officer.id,
-            creator_id=admin.id
+            text_contents=text_contents, officer_id=officer.id, creator_id=admin.id
         )
 
         rv = client.post(
-            url_for('main.description_api', officer_id=officer.id),
+            url_for("main.description_api", officer_id=officer.id),
             data=form.data,
-            follow_redirects=True
+            follow_redirects=True,
         )
 
-        assert rv.status_code == 200
-        assert 'created' in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert "created" in rv.data.decode(ENCODING_UTF_8)
 
-        created_description = Description.query.filter_by(text_contents=text_contents).first()
+        created_description = Description.query.filter_by(
+            text_contents=text_contents
+        ).first()
         assert created_description is not None
         assert created_description.date_created is not None
 
@@ -96,24 +103,24 @@ def test_acs_can_create_descriptions(mockdata, client, session):
     with current_app.test_request_context():
         login_ac(client)
         officer = Officer.query.first()
-        description = 'A description'
-        ac = User.query.filter_by(email='raq929@example.org').first()
+        description = "A description"
+        ac = User.query.filter_by(email="raq929@example.org").first()
         form = TextForm(
-            text_contents=description,
-            officer_id=officer.id,
-            creator_id=ac.id
+            text_contents=description, officer_id=officer.id, creator_id=ac.id
         )
 
         rv = client.post(
-            url_for('main.description_api', officer_id=officer.id),
+            url_for("main.description_api", officer_id=officer.id),
             data=form.data,
-            follow_redirects=True
+            follow_redirects=True,
         )
 
-        assert rv.status_code == 200
-        assert 'created' in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert "created" in rv.data.decode(ENCODING_UTF_8)
 
-        created_description = Description.query.filter_by(text_contents=description).first()
+        created_description = Description.query.filter_by(
+            text_contents=description
+        ).first()
         assert created_description is not None
         assert created_description.date_created is not None
 
@@ -122,8 +129,8 @@ def test_admins_can_edit_descriptions(mockdata, client, session):
     with current_app.test_request_context():
         login_admin(client)
         officer = Officer.query.first()
-        old_description = 'meow'
-        new_description = 'I can haz editing descriptionz'
+        old_description = "meow"
+        new_description = "I can haz editing descriptionz"
         original_date = datetime.now()
         description = Description(
             text_contents=old_description,
@@ -140,12 +147,15 @@ def test_admins_can_edit_descriptions(mockdata, client, session):
         )
 
         rv = client.post(
-            url_for('main.description_api', officer_id=officer.id, obj_id=description.id) + '/edit',
+            url_for(
+                "main.description_api", officer_id=officer.id, obj_id=description.id
+            )
+            + "/edit",
             data=form.data,
-            follow_redirects=True
+            follow_redirects=True,
         )
-        assert rv.status_code == 200
-        assert 'updated' in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert "updated" in rv.data.decode(ENCODING_UTF_8)
 
         assert description.text_contents == new_description
         assert description.date_updated > original_date
@@ -154,10 +164,10 @@ def test_admins_can_edit_descriptions(mockdata, client, session):
 def test_ac_can_edit_their_descriptions_in_their_department(mockdata, client, session):
     with current_app.test_request_context():
         login_ac(client)
-        ac = User.query.filter_by(email='raq929@example.org').first()
+        ac = User.query.filter_by(email="raq929@example.org").first()
         officer = Officer.query.filter_by(department_id=AC_DEPT).first()
-        old_description = 'meow'
-        new_description = 'I can haz editing descriptionz'
+        old_description = "meow"
+        new_description = "I can haz editing descriptionz"
         original_date = datetime.now()
         description = Description(
             text_contents=old_description,
@@ -174,12 +184,15 @@ def test_ac_can_edit_their_descriptions_in_their_department(mockdata, client, se
         )
 
         rv = client.post(
-            url_for('main.description_api', officer_id=officer.id, obj_id=description.id) + '/edit',
+            url_for(
+                "main.description_api", officer_id=officer.id, obj_id=description.id
+            )
+            + "/edit",
             data=form.data,
-            follow_redirects=True
+            follow_redirects=True,
         )
-        assert rv.status_code == 200
-        assert 'updated' in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert "updated" in rv.data.decode(ENCODING_UTF_8)
 
         assert description.text_contents == new_description
         assert description.date_updated > original_date
@@ -188,10 +201,10 @@ def test_ac_can_edit_their_descriptions_in_their_department(mockdata, client, se
 def test_ac_can_edit_others_descriptions(mockdata, client, session):
     with current_app.test_request_context():
         login_ac(client)
-        ac = User.query.filter_by(email='raq929@example.org').first()
+        ac = User.query.filter_by(email="raq929@example.org").first()
         officer = Officer.query.filter_by(department_id=AC_DEPT).first()
-        old_description = 'meow'
-        new_description = 'I can haz editing descriptionz'
+        old_description = "meow"
+        new_description = "I can haz editing descriptionz"
         original_date = datetime.now()
         description = Description(
             text_contents=old_description,
@@ -208,12 +221,15 @@ def test_ac_can_edit_others_descriptions(mockdata, client, session):
         )
 
         rv = client.post(
-            url_for('main.description_api', officer_id=officer.id, obj_id=description.id) + '/edit',
+            url_for(
+                "main.description_api", officer_id=officer.id, obj_id=description.id
+            )
+            + "/edit",
             data=form.data,
-            follow_redirects=True
+            follow_redirects=True,
         )
-        assert rv.status_code == 200
-        assert 'updated' in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert "updated" in rv.data.decode(ENCODING_UTF_8)
 
         assert description.text_contents == new_description
         assert description.date_updated > original_date
@@ -223,10 +239,12 @@ def test_ac_cannot_edit_descriptions_not_in_their_department(mockdata, client, s
     with current_app.test_request_context():
         login_ac(client)
 
-        officer = Officer.query.except_(Officer.query.filter_by(department_id=AC_DEPT)).first()
-        ac = User.query.filter_by(email='raq929@example.org').first()
-        old_description = 'meow'
-        new_description = 'I can haz editing descriptionz'
+        officer = Officer.query.except_(
+            Officer.query.filter_by(department_id=AC_DEPT)
+        ).first()
+        ac = User.query.filter_by(email="raq929@example.org").first()
+        old_description = "meow"
+        new_description = "I can haz editing descriptionz"
         original_date = datetime.now()
         description = Description(
             text_contents=old_description,
@@ -243,11 +261,14 @@ def test_ac_cannot_edit_descriptions_not_in_their_department(mockdata, client, s
         )
 
         rv = client.post(
-            url_for('main.description_api', officer_id=officer.id, obj_id=description.id) + '/edit',
+            url_for(
+                "main.description_api", officer_id=officer.id, obj_id=description.id
+            )
+            + "/edit",
             data=form.data,
-            follow_redirects=True
+            follow_redirects=True,
         )
-        assert rv.status_code == 403
+        assert rv.status_code == HTTPStatus.FORBIDDEN
 
 
 def test_admins_can_delete_descriptions(mockdata, client, session):
@@ -256,21 +277,28 @@ def test_admins_can_delete_descriptions(mockdata, client, session):
         description = Description.query.first()
         description_id = description.id
         rv = client.post(
-            url_for('main.description_api', officer_id=description.officer_id, obj_id=description_id) + '/delete',
-            follow_redirects=True
+            url_for(
+                "main.description_api",
+                officer_id=description.officer_id,
+                obj_id=description_id,
+            )
+            + "/delete",
+            follow_redirects=True,
         )
-        assert rv.status_code == 200
+        assert rv.status_code == HTTPStatus.OK
         deleted = Description.query.get(description_id)
         assert deleted is None
 
 
-def test_acs_can_delete_their_descriptions_in_their_department(mockdata, client, session):
+def test_acs_can_delete_their_descriptions_in_their_department(
+    mockdata, client, session
+):
     with current_app.test_request_context():
         login_ac(client)
-        ac = User.query.filter_by(email='raq929@example.org').first()
+        ac = User.query.filter_by(email="raq929@example.org").first()
         officer = Officer.query.filter_by(department_id=AC_DEPT).first()
         description = Description(
-            text_contents='Hello',
+            text_contents="Hello",
             officer_id=officer.id,
             creator_id=ac.id,
             date_created=datetime.now(),
@@ -280,20 +308,27 @@ def test_acs_can_delete_their_descriptions_in_their_department(mockdata, client,
         db.session.commit()
         description_id = description.id
         rv = client.post(
-            url_for('main.description_api', officer_id=officer.id, obj_id=description.id) + '/delete',
-            follow_redirects=True
+            url_for(
+                "main.description_api", officer_id=officer.id, obj_id=description.id
+            )
+            + "/delete",
+            follow_redirects=True,
         )
-        assert rv.status_code == 200
+        assert rv.status_code == HTTPStatus.OK
         deleted = Description.query.get(description_id)
         assert deleted is None
 
 
-def test_acs_cannot_delete_descriptions_not_in_their_department(mockdata, client, session):
+def test_acs_cannot_delete_descriptions_not_in_their_department(
+    mockdata, client, session
+):
     with current_app.test_request_context():
         login_ac(client)
-        officer = Officer.query.except_(Officer.query.filter_by(department_id=AC_DEPT)).first()
+        officer = Officer.query.except_(
+            Officer.query.filter_by(department_id=AC_DEPT)
+        ).first()
         description = Description(
-            text_contents='Hello',
+            text_contents="Hello",
             officer_id=officer.id,
             creator_id=2,
             date_created=datetime.now(),
@@ -303,11 +338,14 @@ def test_acs_cannot_delete_descriptions_not_in_their_department(mockdata, client
         db.session.commit()
         description_id = description.id
         rv = client.post(
-            url_for('main.description_api', officer_id=officer.id, obj_id=description.id) + '/delete',
-            follow_redirects=True
+            url_for(
+                "main.description_api", officer_id=officer.id, obj_id=description.id
+            )
+            + "/delete",
+            follow_redirects=True,
         )
 
-        assert rv.status_code == 403
+        assert rv.status_code == HTTPStatus.FORBIDDEN
         not_deleted = Description.query.get(description_id)
         assert not_deleted is not None
 
@@ -316,9 +354,9 @@ def test_acs_can_get_edit_form_for_their_dept(mockdata, client, session):
     with current_app.test_request_context():
         login_ac(client)
         officer = Officer.query.filter_by(department_id=AC_DEPT).first()
-        ac = User.query.filter_by(email='raq929@example.org').first()
+        ac = User.query.filter_by(email="raq929@example.org").first()
         description = Description(
-            text_contents='Hello',
+            text_contents="Hello",
             officer_id=officer.id,
             creator_id=ac.id,
             date_created=datetime.now(),
@@ -327,20 +365,23 @@ def test_acs_can_get_edit_form_for_their_dept(mockdata, client, session):
         db.session.add(description)
         db.session.commit()
         rv = client.get(
-            url_for('main.description_api', obj_id=description.id, officer_id=officer.id) + '/edit',
-            follow_redirects=True
+            url_for(
+                "main.description_api", obj_id=description.id, officer_id=officer.id
+            )
+            + "/edit",
+            follow_redirects=True,
         )
-        assert rv.status_code == 200
-        assert 'Update' in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert "Update" in rv.data.decode(ENCODING_UTF_8)
 
 
 def test_acs_can_get_others_edit_form(mockdata, client, session):
     with current_app.test_request_context():
         login_ac(client)
         officer = Officer.query.filter_by(department_id=AC_DEPT).first()
-        ac = User.query.filter_by(email='raq929@example.org').first()
+        ac = User.query.filter_by(email="raq929@example.org").first()
         description = Description(
-            text_contents='Hello',
+            text_contents="Hello",
             officer_id=officer.id,
             creator_id=ac.id - 1,
             date_created=datetime.now(),
@@ -349,19 +390,24 @@ def test_acs_can_get_others_edit_form(mockdata, client, session):
         db.session.add(description)
         db.session.commit()
         rv = client.get(
-            url_for('main.description_api', obj_id=description.id, officer_id=officer.id) + '/edit',
-            follow_redirects=True
+            url_for(
+                "main.description_api", obj_id=description.id, officer_id=officer.id
+            )
+            + "/edit",
+            follow_redirects=True,
         )
-        assert rv.status_code == 200
-        assert 'Update' in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert "Update" in rv.data.decode(ENCODING_UTF_8)
 
 
 def test_acs_cannot_get_edit_form_for_their_non_dept(mockdata, client, session):
     with current_app.test_request_context():
         login_ac(client)
-        officer = Officer.query.except_(Officer.query.filter_by(department_id=AC_DEPT)).first()
+        officer = Officer.query.except_(
+            Officer.query.filter_by(department_id=AC_DEPT)
+        ).first()
         description = Description(
-            text_contents='Hello',
+            text_contents="Hello",
             officer_id=officer.id,
             creator_id=2,
             date_created=datetime.now(),
@@ -370,16 +416,19 @@ def test_acs_cannot_get_edit_form_for_their_non_dept(mockdata, client, session):
         db.session.add(description)
         db.session.commit()
         rv = client.get(
-            url_for('main.description_api', obj_id=description.id, officer_id=officer.id) + '/edit',
-            follow_redirects=True
+            url_for(
+                "main.description_api", obj_id=description.id, officer_id=officer.id
+            )
+            + "/edit",
+            follow_redirects=True,
         )
-        assert rv.status_code == 403
+        assert rv.status_code == HTTPStatus.FORBIDDEN
 
 
 def test_users_can_see_descriptions(mockdata, client, session):
     with current_app.test_request_context():
         officer = Officer.query.first()
-        text_contents = 'You can see me'
+        text_contents = "You can see me"
         description = Description(
             text_contents=text_contents,
             officer_id=officer.id,
@@ -390,20 +439,20 @@ def test_users_can_see_descriptions(mockdata, client, session):
         db.session.add(description)
         db.session.commit()
         rv = client.get(
-            url_for('main.officer_profile', officer_id=officer.id),
-            follow_redirects=True
+            url_for("main.officer_profile", officer_id=officer.id),
+            follow_redirects=True,
         )
         # ensures we're looking for a description that exists
         assert description in officer.descriptions
-        assert rv.status_code == 200
-        assert text_contents in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert text_contents in rv.data.decode(ENCODING_UTF_8)
 
 
 def test_admins_can_see_descriptions(mockdata, client, session):
     with current_app.test_request_context():
         login_admin(client)
         officer = Officer.query.first()
-        text_contents = 'Kittens see everything'
+        text_contents = "Kittens see everything"
         description = Description(
             text_contents=text_contents,
             officer_id=officer.id,
@@ -414,19 +463,19 @@ def test_admins_can_see_descriptions(mockdata, client, session):
         db.session.add(description)
         db.session.commit()
         rv = client.get(
-            url_for('main.officer_profile', officer_id=officer.id),
-            follow_redirects=True
+            url_for("main.officer_profile", officer_id=officer.id),
+            follow_redirects=True,
         )
         assert description in officer.descriptions
-        assert rv.status_code == 200
-        assert text_contents in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert text_contents in rv.data.decode(ENCODING_UTF_8)
 
 
 def test_acs_can_see_descriptions_in_their_department(mockdata, client, session):
     with current_app.test_request_context():
         login_ac(client)
         officer = Officer.query.filter_by(department_id=AC_DEPT).first()
-        text_contents = 'I can haz descriptionz'
+        text_contents = "I can haz descriptionz"
         description = Description(
             text_contents=text_contents,
             officer_id=officer.id,
@@ -437,33 +486,63 @@ def test_acs_can_see_descriptions_in_their_department(mockdata, client, session)
         db.session.add(description)
         db.session.commit()
         rv = client.get(
-            url_for('main.officer_profile', officer_id=officer.id),
-            follow_redirects=True
+            url_for("main.officer_profile", officer_id=officer.id),
+            follow_redirects=True,
         )
         # ensures we're looking for a description that exists
         assert description in officer.descriptions
-        assert rv.status_code == 200
-        assert text_contents in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert text_contents in rv.data.decode(ENCODING_UTF_8)
 
 
 def test_acs_can_see_descriptions_not_in_their_department(mockdata, client, session):
     with current_app.test_request_context():
-        officer = Officer.query.except_(Officer.query.filter_by(department_id=AC_DEPT)).first()
-        text_contents = 'Hello it me'
+        officer = Officer.query.except_(
+            Officer.query.filter_by(department_id=AC_DEPT)
+        ).first()
+        login_ac(client)
+        creator = User.query.get(1)
+        text_contents = "Hello it me"
         description = Description(
             text_contents=text_contents,
             officer_id=officer.id,
-            creator_id=1,
+            creator_id=creator.id,
             date_created=datetime.now(),
             date_updated=datetime.now(),
         )
         db.session.add(description)
         db.session.commit()
         rv = client.get(
-            url_for('main.officer_profile', officer_id=officer.id),
-            follow_redirects=True
+            url_for("main.officer_profile", officer_id=officer.id),
+            follow_redirects=True,
         )
         # ensures we're looking for a description that exists
+        response_text = rv.data.decode(ENCODING_UTF_8)
         assert description in officer.descriptions
-        assert rv.status_code == 200
-        assert text_contents in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert text_contents in response_text
+        assert creator.username in response_text
+
+
+def test_anonymous_users_cannot_see_description_creators(mockdata, client, session):
+    with current_app.test_request_context():
+        officer = Officer.query.first()
+        ac = User.query.filter_by(email="raq929@example.org").first()
+        text_contents = "All we have is each other"
+        description = Description(
+            text_contents=text_contents,
+            officer_id=officer.id,
+            creator_id=ac.id,
+            date_created=datetime.now(),
+            date_updated=datetime.now(),
+        )
+        db.session.add(description)
+        db.session.commit()
+
+        rv = client.get(
+            url_for("main.officer_profile", officer_id=officer.id),
+            follow_redirects=True,
+        )
+        assert description in officer.descriptions
+        assert rv.status_code == HTTPStatus.OK
+        assert ac.username not in rv.data.decode(ENCODING_UTF_8)
