@@ -1,88 +1,137 @@
 # Routing and view tests
+from datetime import date, datetime, time
+from http import HTTPStatus
+
 import pytest
-from datetime import datetime, date, time
-from flask import url_for, current_app
+from flask import current_app, url_for
 from sqlalchemy.orm import joinedload
+
+from OpenOversight.app.main.forms import (
+    IncidentForm,
+    LicensePlateForm,
+    LinkForm,
+    LocationForm,
+    OOIdForm,
+)
+from OpenOversight.app.models import Department, Incident, Officer
+from OpenOversight.app.utils.constants import ENCODING_UTF_8
 from OpenOversight.tests.conftest import AC_DEPT
-from .route_helpers import login_user, login_admin, login_ac, process_form_data
+
+from .route_helpers import login_ac, login_admin, login_user, process_form_data
 
 
-from OpenOversight.app.main.forms import IncidentForm, LocationForm, LinkForm, LicensePlateForm, OOIdForm
-from OpenOversight.app.models import Incident, Officer, Department
-
-
-@pytest.mark.parametrize("route", [
-    ('/incidents/'),
-    ('/incidents/1'),
-    ('/incidents/?department_id=1')
-])
+@pytest.mark.parametrize(
+    "route", ["/incidents/", "/incidents/1", "/incidents/?department_id=1"]
+)
 def test_routes_ok(route, client, mockdata):
     rv = client.get(route)
-    assert rv.status_code == 200
+    assert rv.status_code == HTTPStatus.OK
 
 
-@pytest.mark.parametrize("route", [
-    ('incidents/1/edit'),
-    ('incidents/new'),
-    ('incidents/1/delete')
-])
+@pytest.mark.parametrize(
+    "route", ["incidents/1/edit", "incidents/new", "incidents/1/delete"]
+)
 def test_route_login_required(route, client, mockdata):
     rv = client.get(route)
-    assert rv.status_code == 302
+    assert rv.status_code == HTTPStatus.FOUND
 
 
-@pytest.mark.parametrize("route", [
-    ('incidents/1/edit'),
-    ('incidents/new'),
-    ('incidents/1/delete')
-])
+@pytest.mark.parametrize(
+    "route", ["incidents/1/edit", "incidents/new", "incidents/1/delete"]
+)
 def test_route_admin_or_required(route, client, mockdata):
     with current_app.test_request_context():
         login_user(client)
         rv = client.get(route)
-        assert rv.status_code == 403
+        assert rv.status_code == HTTPStatus.FORBIDDEN
 
 
-def test_admins_can_create_basic_incidents(mockdata, client, session):
+@pytest.mark.parametrize(
+    "report_number",
+    [
+        # Ensure different report number formats are accepted
+        "42",
+        "My-Special-Case",
+        "PPP Case 92",
+    ],
+)
+def test_admins_can_create_basic_incidents(report_number, mockdata, client, session):
     with current_app.test_request_context():
         login_admin(client)
         date = datetime(2000, 5, 25, 1, 45)
-        report_number = '42'
 
         address_form = LocationForm(
-            street_name='AAAAA',
-            cross_street1='BBBBB',
-            city='FFFFF',
-            state='IA',
-            zip_code='03435'
+            street_name="AAAAA",
+            cross_street1="BBBBB",
+            city="FFFFF",
+            state="IA",
+            zip_code="03435",
         )
         # These have to have a dropdown selected because if not, an empty Unicode string is sent, which does not mach the '' selector.
-        link_form = LinkForm(link_type='video')
-        license_plates_form = LicensePlateForm(state='AZ')
+        link_form = LinkForm(link_type="video")
+        license_plates_form = LicensePlateForm(state="AZ")
         form = IncidentForm(
             date_field=str(date.date()),
             time_field=str(date.time()),
             report_number=report_number,
-            description='Something happened',
-            department='1',
+            description="Something happened",
+            department="1",
             address=address_form.data,
             links=[link_form.data],
             license_plates=[license_plates_form.data],
-            officers=[]
-
+            officers=[],
         )
         data = process_form_data(form.data)
 
         rv = client.post(
-            url_for('main.incident_api') + 'new',
-            data=data,
-            follow_redirects=True
+            url_for("main.incident_api") + "new", data=data, follow_redirects=True
         )
-        assert rv.status_code == 200
-        assert 'created' in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert "created" in rv.data.decode(ENCODING_UTF_8)
 
         inc = Incident.query.filter_by(date=date.date()).first()
         assert inc is not None
+
+
+def test_admins_cannot_create_incident_with_invalid_report_number(
+    mockdata, client, session
+):
+    with current_app.test_request_context():
+        login_admin(client)
+        date = datetime(2000, 5, 25, 1, 45)
+        report_number = "Will Not Work! #45"
+
+        address_form = LocationForm(
+            street_name="AAAAA",
+            cross_street1="BBBBB",
+            city="FFFFF",
+            state="IA",
+            zip_code="03435",
+        )
+        # These have to have a dropdown selected because if not, an empty Unicode string is sent, which does not mach the '' selector.
+        link_form = LinkForm(link_type="video")
+        license_plates_form = LicensePlateForm(state="AZ")
+        form = IncidentForm(
+            date_field=str(date.date()),
+            time_field=str(date.time()),
+            report_number=report_number,
+            description="Something happened",
+            department="1",
+            address=address_form.data,
+            links=[link_form.data],
+            license_plates=[license_plates_form.data],
+            officers=[],
+        )
+        data = process_form_data(form.data)
+
+        rv = client.post(
+            url_for("main.incident_api") + "new", data=data, follow_redirects=True
+        )
+
+        assert rv.status_code == HTTPStatus.OK
+        assert "Report cannot contain special characters" in rv.data.decode(
+            ENCODING_UTF_8
+        )
 
 
 def test_admins_can_edit_incident_date_and_address(mockdata, client, session):
@@ -91,20 +140,26 @@ def test_admins_can_edit_incident_date_and_address(mockdata, client, session):
         inc = Incident.query.options(
             joinedload(Incident.links),
             joinedload(Incident.license_plates),
-            joinedload(Incident.officers)).first()
+            joinedload(Incident.officers),
+        ).first()
         inc_id = inc.id
         new_date = date(2017, 6, 25)
         new_time = time(1, 45)
-        street_name = 'Newest St'
+        street_name = "Newest St"
         address_form = LocationForm(
             street_name=street_name,
-            cross_street1='Your St',
-            city='Boston',
-            state='NH',
-            zip_code='03435'
+            cross_street1="Your St",
+            city="Boston",
+            state="NH",
+            zip_code="03435",
         )
-        links_forms = [LinkForm(url=link.url, link_type=link.link_type).data for link in inc.links]
-        license_plates_forms = [LicensePlateForm(number=lp.number, state=lp.state).data for lp in inc.license_plates]
+        links_forms = [
+            LinkForm(url=link.url, link_type=link.link_type).data for link in inc.links
+        ]
+        license_plates_forms = [
+            LicensePlateForm(number=lp.number, state=lp.state).data
+            for lp in inc.license_plates
+        ]
         ooid_forms = [OOIdForm(ooid=officer.id) for officer in inc.officers]
 
         form = IncidentForm(
@@ -112,21 +167,21 @@ def test_admins_can_edit_incident_date_and_address(mockdata, client, session):
             time_field=str(new_time),
             report_number=inc.report_number,
             description=inc.description,
-            department='1',
+            department="1",
             address=address_form.data,
             links=links_forms,
             license_plates=license_plates_forms,
-            officers=ooid_forms
+            officers=ooid_forms,
         )
         data = process_form_data(form.data)
 
         rv = client.post(
-            url_for('main.incident_api', obj_id=inc.id) + '/edit',
+            url_for("main.incident_api", obj_id=inc.id) + "/edit",
             data=data,
-            follow_redirects=True
+            follow_redirects=True,
         )
-        assert rv.status_code == 200
-        assert 'successfully updated' in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert "successfully updated" in rv.data.decode(ENCODING_UTF_8)
         updated = Incident.query.get(inc_id)
         assert updated.date == new_date
         assert updated.time == new_time
@@ -139,7 +194,8 @@ def test_admins_can_edit_incident_links_and_licenses(mockdata, client, session):
         inc = Incident.query.options(
             joinedload(Incident.links),
             joinedload(Incident.license_plates),
-            joinedload(Incident.officers)).first()
+            joinedload(Incident.officers),
+        ).first()
 
         address_form = LocationForm(
             street_name=inc.address.street_name,
@@ -147,15 +203,17 @@ def test_admins_can_edit_incident_links_and_licenses(mockdata, client, session):
             cross_street2=inc.address.cross_street2,
             city=inc.address.city,
             state=inc.address.state,
-            zip_code=inc.address.zip_code
+            zip_code=inc.address.zip_code,
         )
         old_links = inc.links
-        old_links_forms = [LinkForm(url=link.url, link_type=link.link_type).data for link in inc.links]
-        new_url = 'http://rachel.com'
-        link_form = LinkForm(url='http://rachel.com', link_type='video')
+        old_links_forms = [
+            LinkForm(url=link.url, link_type=link.link_type).data for link in inc.links
+        ]
+        new_url = "http://rachel.com"
+        link_form = LinkForm(url="http://rachel.com", link_type="video")
         old_license_plates = inc.license_plates
-        new_number = '453893'
-        license_plates_form = LicensePlateForm(number=new_number, state='IA')
+        new_number = "453893"
+        license_plates_form = LicensePlateForm(number=new_number, state="IA")
         ooid_forms = [OOIdForm(ooid=officer.id) for officer in inc.officers]
 
         form = IncidentForm(
@@ -163,21 +221,21 @@ def test_admins_can_edit_incident_links_and_licenses(mockdata, client, session):
             time_field=str(inc.time),
             report_number=inc.report_number,
             description=inc.description,
-            department='1',
+            department="1",
             address=address_form.data,
             links=old_links_forms + [link_form.data],
             license_plates=[license_plates_form.data],
-            officers=ooid_forms
+            officers=ooid_forms,
         )
         data = process_form_data(form.data)
 
         rv = client.post(
-            url_for('main.incident_api', obj_id=inc.id) + '/edit',
+            url_for("main.incident_api", obj_id=inc.id) + "/edit",
             data=data,
-            follow_redirects=True
+            follow_redirects=True,
         )
-        assert rv.status_code == 200
-        assert 'successfully updated' in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert "successfully updated" in rv.data.decode(ENCODING_UTF_8)
         # old links are still there
         for link in old_links:
             assert link in inc.links
@@ -194,7 +252,8 @@ def test_admins_cannot_make_ancient_incidents(mockdata, client, session):
         inc = Incident.query.options(
             joinedload(Incident.links),
             joinedload(Incident.license_plates),
-            joinedload(Incident.officers)).first()
+            joinedload(Incident.officers),
+        ).first()
 
         address_form = LocationForm(
             street_name=inc.address.street_name,
@@ -202,7 +261,7 @@ def test_admins_cannot_make_ancient_incidents(mockdata, client, session):
             cross_street2=inc.address.cross_street2,
             city=inc.address.city,
             state=inc.address.state,
-            zip_code=inc.address.zip_code
+            zip_code=inc.address.zip_code,
         )
         ooid_forms = [OOIdForm(ooid=officer.id) for officer in inc.officers]
 
@@ -211,104 +270,100 @@ def test_admins_cannot_make_ancient_incidents(mockdata, client, session):
             time_field=str(inc.time),
             report_number=inc.report_number,
             description=inc.description,
-            department='1',
+            department="1",
             address=address_form.data,
-            officers=ooid_forms
+            officers=ooid_forms,
         )
         data = process_form_data(form.data)
 
         rv = client.post(
-            url_for('main.incident_api', obj_id=inc.id) + '/edit',
+            url_for("main.incident_api", obj_id=inc.id) + "/edit",
             data=data,
-            follow_redirects=True
+            follow_redirects=True,
         )
-        assert rv.status_code == 200
-        assert 'Incidents prior to 1900 not allowed.' in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert "Incidents prior to 1900 not allowed." in rv.data.decode(ENCODING_UTF_8)
 
 
 def test_admins_cannot_make_incidents_without_state(mockdata, client, session):
     with current_app.test_request_context():
         login_admin(client)
         date = datetime(2000, 5, 25, 1, 45)
-        report_number = '42'
+        report_number = "42"
 
         address_form = LocationForm(
-            street_name='AAAAA',
-            cross_street1='BBBBB',
-            city='FFFFF',
-            state='',
-            zip_code='03435'
+            street_name="AAAAA",
+            cross_street1="BBBBB",
+            city="FFFFF",
+            state="",
+            zip_code="03435",
         )
-        ooid_forms = [OOIdForm(ooid=officer.id)
-                      for officer in Officer.query.all()[:5]]
+        ooid_forms = [OOIdForm(ooid=officer.id) for officer in Officer.query.all()[:5]]
 
         form = IncidentForm(
             date_field=str(date.date()),
             time_field=str(date.time()),
             report_number=report_number,
-            description='Something happened',
-            department='1',
+            description="Something happened",
+            department="1",
             address=address_form.data,
-            officers=ooid_forms
+            officers=ooid_forms,
         )
         data = process_form_data(form.data)
 
         incident_count_before = Incident.query.count()
         rv = client.post(
-            url_for('main.incident_api') + 'new',
-            data=data,
-            follow_redirects=True
+            url_for("main.incident_api") + "new", data=data, follow_redirects=True
         )
-        assert rv.status_code == 200
-        assert 'Must select a state.' in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert "Must select a state." in rv.data.decode(ENCODING_UTF_8)
         assert incident_count_before == Incident.query.count()
 
 
-def test_admins_cannot_make_incidents_with_multiple_validation_errors(mockdata, client, session):
+def test_admins_cannot_make_incidents_with_multiple_validation_errors(
+    mockdata, client, session
+):
     with current_app.test_request_context():
         login_admin(client)
         date = datetime(2000, 5, 25, 1, 45)
-        report_number = '42'
+        report_number = "42"
 
         address_form = LocationForm(
-            street_name='AAAAA',
-            cross_street1='BBBBB',
+            street_name="AAAAA",
+            cross_street1="BBBBB",
             # no city given => 'This field is required.'
-            city='',
-            state='NY',
+            city="",
+            state="NY",
             # invalid ZIP code => 'Zip codes must have 5 digits.'
-            zip_code='0343'
+            zip_code="0343",
         )
 
         # license plate number given, but no state selected => 'Must also select a state.'
-        license_plate_form = LicensePlateForm(number='ABCDE', state='')
-        ooid_forms = [OOIdForm(ooid=officer.id)
-                      for officer in Officer.query.all()[:5]]
+        license_plate_form = LicensePlateForm(number="ABCDE", state="")
+        ooid_forms = [OOIdForm(ooid=officer.id) for officer in Officer.query.all()[:5]]
 
         form = IncidentForm(
             # no date given => 'This field is required.'
-            date_field='',
+            date_field="",
             time_field=str(date.time()),
             report_number=report_number,
-            description='Something happened',
+            description="Something happened",
             # invalid department id => 'This field is required.'
-            department='-1',
+            department="-1",
             address=address_form.data,
             license_plates=[license_plate_form.data],
-            officers=ooid_forms
+            officers=ooid_forms,
         )
         data = process_form_data(form.data)
 
         incident_count_before = Incident.query.count()
         rv = client.post(
-            url_for('main.incident_api') + 'new',
-            data=data,
-            follow_redirects=True
+            url_for("main.incident_api") + "new", data=data, follow_redirects=True
         )
-        assert rv.status_code == 200
-        assert 'Must also select a state.' in rv.data.decode('utf-8')
-        assert 'Zip codes must have 5 digits.' in rv.data.decode('utf-8')
-        assert rv.data.decode('utf-8').count('This field is required.') >= 3
+        assert rv.status_code == HTTPStatus.OK
+        assert "Must also select a state." in rv.data.decode(ENCODING_UTF_8)
+        assert "Zip codes must have 5 digits." in rv.data.decode(ENCODING_UTF_8)
+        assert rv.data.decode(ENCODING_UTF_8).count("This field is required.") >= 3
         assert incident_count_before == Incident.query.count()
 
 
@@ -318,7 +373,8 @@ def test_admins_can_edit_incident_officers(mockdata, client, session):
         inc = Incident.query.options(
             joinedload(Incident.links),
             joinedload(Incident.license_plates),
-            joinedload(Incident.officers)).first()
+            joinedload(Incident.officers),
+        ).first()
 
         address_form = LocationForm(
             street_name=inc.address.street_name,
@@ -326,16 +382,23 @@ def test_admins_can_edit_incident_officers(mockdata, client, session):
             cross_street2=inc.address.cross_street2,
             city=inc.address.city,
             state=inc.address.state,
-            zip_code=inc.address.zip_code
+            zip_code=inc.address.zip_code,
         )
-        links_forms = [LinkForm(url=link.url, link_type=link.link_type).data for link in inc.links]
-        license_plates_forms = [LicensePlateForm(number=lp.number, state=lp.state).data for lp in inc.license_plates]
+        links_forms = [
+            LinkForm(url=link.url, link_type=link.link_type).data for link in inc.links
+        ]
+        license_plates_forms = [
+            LicensePlateForm(number=lp.number, state=lp.state).data
+            for lp in inc.license_plates
+        ]
 
         old_officers = inc.officers
         old_officer_ids = [officer.id for officer in inc.officers]
         old_ooid_forms = [OOIdForm(oo_id=the_id) for the_id in old_officer_ids]
         # get a new officer that is different from the old officers
-        new_officer = Officer.query.except_(Officer.query.filter(Officer.id.in_(old_officer_ids))).first()
+        new_officer = Officer.query.except_(
+            Officer.query.filter(Officer.id.in_(old_officer_ids))
+        ).first()
         new_ooid_form = OOIdForm(oo_id=new_officer.id)
 
         form = IncidentForm(
@@ -343,21 +406,21 @@ def test_admins_can_edit_incident_officers(mockdata, client, session):
             time_field=str(inc.time),
             report_number=inc.report_number,
             description=inc.description,
-            department='1',
+            department="1",
             address=address_form.data,
             links=links_forms,
             license_plates=license_plates_forms,
-            officers=old_ooid_forms + [new_ooid_form]
+            officers=old_ooid_forms + [new_ooid_form],
         )
         data = process_form_data(form.data)
 
         rv = client.post(
-            url_for('main.incident_api', obj_id=inc.id) + '/edit',
+            url_for("main.incident_api", obj_id=inc.id) + "/edit",
             data=data,
-            follow_redirects=True
+            follow_redirects=True,
         )
-        assert rv.status_code == 200
-        assert 'successfully updated' in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert "successfully updated" in rv.data.decode(ENCODING_UTF_8)
         for officer in old_officers:
             assert officer in inc.officers
         assert new_officer.id in [off.id for off in inc.officers]
@@ -369,7 +432,8 @@ def test_admins_cannot_edit_nonexisting_officers(mockdata, client, session):
         inc = Incident.query.options(
             joinedload(Incident.links),
             joinedload(Incident.license_plates),
-            joinedload(Incident.officers)).first()
+            joinedload(Incident.officers),
+        ).first()
 
         address_form = LocationForm(
             street_name=inc.address.street_name,
@@ -377,10 +441,15 @@ def test_admins_cannot_edit_nonexisting_officers(mockdata, client, session):
             cross_street2=inc.address.cross_street2,
             city=inc.address.city,
             state=inc.address.state,
-            zip_code=inc.address.zip_code
+            zip_code=inc.address.zip_code,
         )
-        links_forms = [LinkForm(url=link.url, link_type=link.link_type).data for link in inc.links]
-        license_plates_forms = [LicensePlateForm(number=lp.number, state=lp.state).data for lp in inc.license_plates]
+        links_forms = [
+            LinkForm(url=link.url, link_type=link.link_type).data for link in inc.links
+        ]
+        license_plates_forms = [
+            LicensePlateForm(number=lp.number, state=lp.state).data
+            for lp in inc.license_plates
+        ]
 
         old_officers = inc.officers
         old_officer_ids = [officer.id for officer in inc.officers]
@@ -393,21 +462,21 @@ def test_admins_cannot_edit_nonexisting_officers(mockdata, client, session):
             time_field=str(inc.time),
             report_number=inc.report_number,
             description=inc.description,
-            department='1',
+            department="1",
             address=address_form.data,
             links=links_forms,
             license_plates=license_plates_forms,
-            officers=old_ooid_forms + [new_ooid_form]
+            officers=old_ooid_forms + [new_ooid_form],
         )
         data = process_form_data(form.data)
 
         rv = client.post(
-            url_for('main.incident_api', obj_id=inc.id) + '/edit',
+            url_for("main.incident_api", obj_id=inc.id) + "/edit",
             data=data,
-            follow_redirects=True
+            follow_redirects=True,
         )
-        assert rv.status_code == 200
-        assert 'Not a valid officer id' in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert "Not a valid officer id" in rv.data.decode(ENCODING_UTF_8)
         for officer in old_officers:
             assert officer in inc.officers
 
@@ -417,16 +486,21 @@ def test_ac_can_edit_incidents_in_their_department(mockdata, client, session):
         login_ac(client)
         inc = Incident.query.filter_by(department_id=AC_DEPT).first()
         new_date = datetime(2017, 6, 25, 1, 45)
-        street_name = 'Newest St'
+        street_name = "Newest St"
         address_form = LocationForm(
             street_name=street_name,
-            cross_street1='Your St',
-            city='Boston',
-            state='NH',
-            zip_code='03435'
+            cross_street1="Your St",
+            city="Boston",
+            state="NH",
+            zip_code="03435",
         )
-        links_forms = [LinkForm(url=link.url, link_type=link.link_type).data for link in inc.links]
-        license_plates_forms = [LicensePlateForm(number=lp.number, state=lp.state).data for lp in inc.license_plates]
+        links_forms = [
+            LinkForm(url=link.url, link_type=link.link_type).data for link in inc.links
+        ]
+        license_plates_forms = [
+            LicensePlateForm(number=lp.number, state=lp.state).data
+            for lp in inc.license_plates
+        ]
         ooid_forms = [OOIdForm(ooid=officer.id) for officer in inc.officers]
 
         form = IncidentForm(
@@ -438,17 +512,17 @@ def test_ac_can_edit_incidents_in_their_department(mockdata, client, session):
             address=address_form.data,
             links=links_forms,
             license_plates=license_plates_forms,
-            officers=ooid_forms
+            officers=ooid_forms,
         )
         data = process_form_data(form.data)
 
         rv = client.post(
-            url_for('main.incident_api', obj_id=inc.id) + '/edit',
+            url_for("main.incident_api", obj_id=inc.id) + "/edit",
             data=data,
-            follow_redirects=True
+            follow_redirects=True,
         )
-        assert rv.status_code == 200
-        assert 'successfully updated' in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert "successfully updated" in rv.data.decode(ENCODING_UTF_8)
         assert inc.date == new_date.date()
         assert inc.time == new_date.time()
         assert inc.address.street_name == street_name
@@ -458,18 +532,25 @@ def test_ac_cannot_edit_incidents_not_in_their_department(mockdata, client, sess
     with current_app.test_request_context():
         login_ac(client)
 
-        inc = Incident.query.except_(Incident.query.filter_by(department_id=AC_DEPT)).first()
+        inc = Incident.query.except_(
+            Incident.query.filter_by(department_id=AC_DEPT)
+        ).first()
         new_date = datetime(2017, 6, 25, 1, 45)
-        street_name = 'Not Allowed St'
+        street_name = "Not Allowed St"
         address_form = LocationForm(
             street_name=street_name,
-            cross_street1='Your St',
-            city='Boston',
-            state='NH',
-            zip_code='03435'
+            cross_street1="Your St",
+            city="Boston",
+            state="NH",
+            zip_code="03435",
         )
-        links_forms = [LinkForm(url=link.url, link_type=link.link_type).data for link in inc.links]
-        license_plates_forms = [LicensePlateForm(number=lp.number, state=lp.state).data for lp in inc.license_plates]
+        links_forms = [
+            LinkForm(url=link.url, link_type=link.link_type).data for link in inc.links
+        ]
+        license_plates_forms = [
+            LicensePlateForm(number=lp.number, state=lp.state).data
+            for lp in inc.license_plates
+        ]
         ooid_forms = [OOIdForm(ooid=officer.id) for officer in inc.officers]
 
         form = IncidentForm(
@@ -481,16 +562,16 @@ def test_ac_cannot_edit_incidents_not_in_their_department(mockdata, client, sess
             address=address_form.data,
             links=links_forms,
             license_plates=license_plates_forms,
-            officers=ooid_forms
+            officers=ooid_forms,
         )
         data = process_form_data(form.data)
 
         rv = client.post(
-            url_for('main.incident_api', obj_id=inc.id) + '/edit',
+            url_for("main.incident_api", obj_id=inc.id) + "/edit",
             data=data,
-            follow_redirects=True
+            follow_redirects=True,
         )
-        assert rv.status_code == 403
+        assert rv.status_code == HTTPStatus.FORBIDDEN
 
 
 def test_admins_can_delete_incidents(mockdata, client, session):
@@ -499,10 +580,10 @@ def test_admins_can_delete_incidents(mockdata, client, session):
         incident = Incident.query.first()
         inc_id = incident.id
         rv = client.post(
-            url_for('main.incident_api', obj_id=inc_id) + '/delete',
-            follow_redirects=True
+            url_for("main.incident_api", obj_id=inc_id) + "/delete",
+            follow_redirects=True,
         )
-        assert rv.status_code == 200
+        assert rv.status_code == HTTPStatus.OK
         deleted = Incident.query.get(inc_id)
         assert deleted is None
 
@@ -513,10 +594,10 @@ def test_acs_can_delete_incidents_in_their_department(mockdata, client, session)
         incident = Incident.query.filter_by(department_id=AC_DEPT).first()
         inc_id = incident.id
         rv = client.post(
-            url_for('main.incident_api', obj_id=inc_id) + '/delete',
-            follow_redirects=True
+            url_for("main.incident_api", obj_id=inc_id) + "/delete",
+            follow_redirects=True,
         )
-        assert rv.status_code == 200
+        assert rv.status_code == HTTPStatus.OK
         deleted = Incident.query.get(inc_id)
         assert deleted is None
 
@@ -524,13 +605,15 @@ def test_acs_can_delete_incidents_in_their_department(mockdata, client, session)
 def test_acs_cannot_delete_incidents_not_in_their_department(mockdata, client, session):
     with current_app.test_request_context():
         login_ac(client)
-        incident = Incident.query.except_(Incident.query.filter_by(department_id=AC_DEPT)).first()
+        incident = Incident.query.except_(
+            Incident.query.filter_by(department_id=AC_DEPT)
+        ).first()
         inc_id = incident.id
         rv = client.post(
-            url_for('main.incident_api', obj_id=inc_id) + '/delete',
-            follow_redirects=True
+            url_for("main.incident_api", obj_id=inc_id) + "/delete",
+            follow_redirects=True,
         )
-        assert rv.status_code == 403
+        assert rv.status_code == HTTPStatus.FORBIDDEN
         not_deleted = Incident.query.get(inc_id)
         assert not_deleted.id is inc_id
 
@@ -540,72 +623,81 @@ def test_acs_can_get_edit_form_for_their_dept(mockdata, client, session):
         login_ac(client)
         incident = Incident.query.filter_by(department_id=AC_DEPT).first()
         rv = client.get(
-            url_for('main.incident_api', obj_id=incident.id) + '/edit',
-            follow_redirects=True
+            url_for("main.incident_api", obj_id=incident.id) + "/edit",
+            follow_redirects=True,
         )
-        assert rv.status_code == 200
-        assert 'Update' in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert "Update" in rv.data.decode(ENCODING_UTF_8)
 
 
 def test_acs_cannot_get_edit_form_for_their_non_dept(mockdata, client, session):
     with current_app.test_request_context():
         login_ac(client)
-        incident = Incident.query.except_(Incident.query.filter_by(department_id=AC_DEPT)).first()
+        incident = Incident.query.except_(
+            Incident.query.filter_by(department_id=AC_DEPT)
+        ).first()
         rv = client.get(
-            url_for('main.incident_api', obj_id=incident.id) + '/edit',
-            follow_redirects=True
+            url_for("main.incident_api", obj_id=incident.id) + "/edit",
+            follow_redirects=True,
         )
-        assert rv.status_code == 403
+        assert rv.status_code == HTTPStatus.FORBIDDEN
 
 
 def test_users_can_view_incidents_by_department(mockdata, client, session):
     with current_app.test_request_context():
         department = Department.query.first()
         department_incidents = Incident.query.filter_by(department_id=department.id)
-        non_department_incidents = Incident.query.except_(Incident.query.filter_by(department_id=department.id))
-        rv = client.get(
-            url_for('main.incident_api', department_id=department.id))
+        non_department_incidents = Incident.query.except_(
+            Incident.query.filter_by(department_id=department.id)
+        )
+        rv = client.get(url_for("main.incident_api", department_id=department.id))
 
         # Requires that report numbers in test data not include other report numbers
         # Tests for report numbers in table formatting, because testing for the raw report number can get false positives due to html encoding
         for incident in department_incidents:
-            assert '<td>{}</td>'.format(incident.report_number) in rv.data.decode('utf-8')
+            assert "<td>{}</td>".format(incident.report_number) in rv.data.decode(
+                ENCODING_UTF_8
+            )
         for incident in non_department_incidents:
-            assert '<td>{}</td>'.format(incident.report_number) not in rv.data.decode('utf-8')
+            assert "<td>{}</td>".format(incident.report_number) not in rv.data.decode(
+                ENCODING_UTF_8
+            )
 
 
 def test_admins_can_see_who_created_incidents(mockdata, client, session):
     with current_app.test_request_context():
         login_admin(client)
-        rv = client.get(url_for('main.incident_api', obj_id=1))
-        assert 'Creator' in rv.data.decode('utf-8')
+        rv = client.get(url_for("main.incident_api", obj_id=1))
+        assert "Creator" in rv.data.decode(ENCODING_UTF_8)
 
 
 def test_acs_cannot_see_who_created_incidents(mockdata, client, session):
     with current_app.test_request_context():
         login_ac(client)
-        rv = client.get(url_for('main.incident_api', obj_id=1))
-        assert 'Creator' not in rv.data.decode('utf-8')
+        rv = client.get(url_for("main.incident_api", obj_id=1))
+        assert "Creator" not in rv.data.decode(ENCODING_UTF_8)
 
 
 def test_users_cannot_see_who_created_incidents(mockdata, client, session):
     with current_app.test_request_context():
         login_ac(client)
-        rv = client.get(url_for('main.incident_api', obj_id=1))
-        assert 'Creator' not in rv.data.decode('utf-8')
+        rv = client.get(url_for("main.incident_api", obj_id=1))
+        assert "Creator" not in rv.data.decode(ENCODING_UTF_8)
 
 
 def test_form_with_officer_id_prepopulates(mockdata, client, session):
     with current_app.test_request_context():
         login_admin(client)
-        officer_id = '1234'
-        rv = client.get(url_for('main.incident_api') + 'new?officer_id={}'.format(officer_id))
-        assert officer_id in rv.data.decode('utf-8')
+        officer_id = "1234"
+        rv = client.get(
+            url_for("main.incident_api") + "new?officer_id={}".format(officer_id)
+        )
+        assert officer_id in rv.data.decode(ENCODING_UTF_8)
 
 
 def test_incident_markdown(mockdata, client, session):
     with current_app.test_request_context():
-        rv = client.get(url_for('main.incident_api'))
+        rv = client.get(url_for("main.incident_api"))
         html = rv.data.decode()
         assert "<h3>A thing happened</h3>" in html
         assert "<p><strong>Markup</strong> description</p>" in html
@@ -617,7 +709,8 @@ def test_admins_cannot_inject_unsafe_html(mockdata, client, session):
         inc = Incident.query.options(
             joinedload(Incident.links),
             joinedload(Incident.license_plates),
-            joinedload(Incident.officers)).first()
+            joinedload(Incident.officers),
+        ).first()
 
         address_form = LocationForm(
             street_name=inc.address.street_name,
@@ -625,10 +718,15 @@ def test_admins_cannot_inject_unsafe_html(mockdata, client, session):
             cross_street2=inc.address.cross_street2,
             city=inc.address.city,
             state=inc.address.state,
-            zip_code=inc.address.zip_code
+            zip_code=inc.address.zip_code,
         )
-        links_forms = [LinkForm(url=link.url, link_type=link.link_type).data for link in inc.links]
-        license_plates_forms = [LicensePlateForm(number=lp.number, state=lp.state).data for lp in inc.license_plates]
+        links_forms = [
+            LinkForm(url=link.url, link_type=link.link_type).data for link in inc.links
+        ]
+        license_plates_forms = [
+            LicensePlateForm(number=lp.number, state=lp.state).data
+            for lp in inc.license_plates
+        ]
 
         old_officer_ids = [officer.id for officer in inc.officers]
         ooid_forms = [OOIdForm(oo_id=the_id) for the_id in old_officer_ids]
@@ -638,20 +736,50 @@ def test_admins_cannot_inject_unsafe_html(mockdata, client, session):
             time_field=str(inc.time),
             report_number=inc.report_number,
             description="<script>alert();</script>",
-            department='1',
+            department="1",
             address=address_form.data,
             links=links_forms,
             license_plates=license_plates_forms,
-            officers=ooid_forms
+            officers=ooid_forms,
         )
         data = process_form_data(form.data)
 
         rv = client.post(
-            url_for('main.incident_api', obj_id=inc.id) + '/edit',
+            url_for("main.incident_api", obj_id=inc.id) + "/edit",
             data=data,
-            follow_redirects=True
+            follow_redirects=True,
         )
-        assert rv.status_code == 200
-        assert 'successfully updated' in rv.data.decode('utf-8')
+        assert rv.status_code == HTTPStatus.OK
+        assert "successfully updated" in rv.data.decode(ENCODING_UTF_8)
         assert "<script>" not in rv.data.decode()
         assert "&lt;script&gt;" in rv.data.decode()
+
+
+@pytest.mark.parametrize(
+    "params,included_report_nums,excluded_report_nums",
+    [
+        ({"department_id": "1"}, ["42"], ["38", "39"]),
+        ({"occurred_before": "2017-12-12"}, ["38", "42"], ["39"]),
+        (
+            {"occurred_after": "2017-12-10", "occurred_before": "2019-01-01"},
+            ["38"],
+            ["39", "42"],
+        ),
+        ({"report_number": "38"}, ["38"], ["42", "39"]),  # Base case
+        ({"report_number": "3"}, ["38", "39"], ["42"]),  # Test inclusive match
+        ({"report_number": "38 "}, ["38"], ["42", "39"]),  # Test trim
+    ],
+)
+def test_users_can_search_incidents(
+    params, included_report_nums, excluded_report_nums, mockdata, client, session
+):
+    with current_app.test_request_context():
+        rv = client.get(url_for("main.incident_api", **params))
+
+        for report_num in included_report_nums:
+            assert "<td>{}</td>".format(report_num) in rv.data.decode(ENCODING_UTF_8)
+
+        for report_num in excluded_report_nums:
+            assert "<td>{}</td>".format(report_num) not in rv.data.decode(
+                ENCODING_UTF_8
+            )
