@@ -5,16 +5,16 @@ import os
 import traceback
 import uuid
 
+import pandas as pd
+import pytest
 from click.testing import CliRunner
 from sqlalchemy.orm.exc import MultipleResultsFound
 
-import pandas as pd
-import pytest
 from OpenOversight.app.commands import (
     add_department,
     add_job_title,
-    bulk_add_officers,
     advanced_csv_import,
+    bulk_add_officers,
     create_officer_from_row,
 )
 from OpenOversight.app.models import (
@@ -27,7 +27,7 @@ from OpenOversight.app.models import (
     Salary,
     Unit,
 )
-from OpenOversight.app.utils import get_officer
+from OpenOversight.app.utils.db import get_officer
 from OpenOversight.tests.conftest import RANK_CHOICES_1, generate_officer
 
 
@@ -103,9 +103,9 @@ def test_add_department__missing_argument(session):
 def test_add_job_title__success(session, department):
     department_id = department.id
 
-    job_title = "Police Officer"
+    job_title = "New Rank"
     is_sworn = True
-    order = 1
+    order = 15
 
     # run command to add job title
     result = run_command_print_output(
@@ -128,14 +128,14 @@ def test_add_job_title__duplicate(session, department):
     job_title = "Police Officer"
     is_sworn = True
     order = 1
-    job = Job(
-        job_title=job_title,
-        is_sworn_officer=is_sworn,
-        order=order,
-        department=department,
+
+    # make sure Job is already included in db
+    assert (
+        Job.query.filter_by(
+            job_title=job_title, is_sworn_officer=True, order=1, department=department
+        ).first()
+        is not None
     )
-    session.add(job)
-    session.commit()
 
     # adding exact same job again via command
     result = run_command_print_output(
@@ -156,14 +156,13 @@ def test_add_job_title__different_departments(session, department):
     job_title = "Police Officer"
     is_sworn = True
     order = 1
-    job = Job(
-        job_title=job_title,
-        is_sworn_officer=is_sworn,
-        order=order,
-        department=department,
+    # make sure Job is already included in db
+    assert (
+        Job.query.filter_by(
+            job_title=job_title, is_sworn_officer=True, order=1, department=department
+        ).first()
+        is not None
     )
-    session.add(job)
-    session.commit()
 
     # adding samme job but for different department
     result = run_command_print_output(
@@ -185,7 +184,7 @@ def test_add_job_title__different_departments(session, department):
 
 
 def test_csv_import_new(csvfile, monkeypatch):
-    monkeypatch.setattr('builtins.input', lambda: "y")
+    monkeypatch.setattr("builtins.input", lambda: "y")
     # Delete all current officers
     Officer.query.delete()
 
@@ -199,7 +198,7 @@ def test_csv_import_new(csvfile, monkeypatch):
 
 
 def test_csv_import_update(csvfile, monkeypatch):
-    monkeypatch.setattr('builtins.input', lambda: "y")
+    monkeypatch.setattr("builtins.input", lambda: "y")
     n_existing = Officer.query.count()
 
     assert n_existing > 0
@@ -212,7 +211,7 @@ def test_csv_import_update(csvfile, monkeypatch):
 
 
 def test_csv_import_idempotence(csvfile, monkeypatch):
-    monkeypatch.setattr('builtins.input', lambda: "y")
+    monkeypatch.setattr("builtins.input", lambda: "y")
     # Delete all current officers
     Officer.query.delete()
 
@@ -283,7 +282,7 @@ def test_csv_changed_static_field(csvfile):
 
 
 def test_csv_new_assignment(csvfile, monkeypatch):
-    monkeypatch.setattr('builtins.input', lambda: "y")
+    monkeypatch.setattr("builtins.input", lambda: "y")
     # Delete all current officers and assignments
     Assignment.query.delete()
     Officer.query.delete()
@@ -324,7 +323,7 @@ def test_csv_new_assignment(csvfile, monkeypatch):
 
 
 def test_csv_new_name(csvfile, monkeypatch):
-    monkeypatch.setattr('builtins.input', lambda: "y")
+    monkeypatch.setattr("builtins.input", lambda: "y")
     df = pd.read_csv(csvfile)
     officer_uid = df.loc[0, "unique_internal_identifier"]
     assert officer_uid
@@ -342,7 +341,7 @@ def test_csv_new_name(csvfile, monkeypatch):
 
 
 def test_csv_new_officer(csvfile, monkeypatch):
-    monkeypatch.setattr('builtins.input', lambda: "y")
+    monkeypatch.setattr("builtins.input", lambda: "y")
     df = pd.read_csv(csvfile)
 
     n_rows = len(df.index)
@@ -352,28 +351,32 @@ def test_csv_new_officer(csvfile, monkeypatch):
     assert n_officers > 0
 
     new_uid = str(uuid.uuid4())
-    new_officer = {  # Must match fields in csvfile
-        "department_id": 1,
-        "unique_internal_identifier": new_uid,
-        "first_name": "FOO",
-        "last_name": "BAR",
-        "middle_initial": None,
-        "suffix": None,
-        "gender": "F",
-        "race": "BLACK",
-        "employment_date": None,
-        "birth_year": None,
-        "star_no": 666,
-        "job_title": "CAPTAIN",
-        "unit": None,
-        "star_date": None,
-        "resign_date": None,
-        "salary": 1.23,
-        "salary_year": 2019,
-        "salary_is_fiscal_year": True,
-        "overtime_pay": 4.56,
-    }
-    df = df.append([new_officer])
+    new_officer = pd.DataFrame.from_dict(
+        [
+            {  # Must match fields in csvfile
+                "department_id": 1,
+                "unique_internal_identifier": new_uid,
+                "first_name": "FOO",
+                "last_name": "BAR",
+                "middle_initial": None,
+                "suffix": None,
+                "gender": "F",
+                "race": "BLACK",
+                "employment_date": None,
+                "birth_year": None,
+                "star_no": 666,
+                "job_title": "CAPTAIN",
+                "unit": None,
+                "star_date": None,
+                "resign_date": None,
+                "salary": 1.23,
+                "salary_year": 2019,
+                "salary_is_fiscal_year": True,
+                "overtime_pay": 4.56,
+            }
+        ]
+    )
+    df = pd.concat([df, new_officer])
     df.to_csv(csvfile)
 
     n_created, n_updated = bulk_add_officers([csvfile], standalone_mode=False)
@@ -387,7 +390,7 @@ def test_csv_new_officer(csvfile, monkeypatch):
 
 
 def test_csv_new_salary(csvfile, monkeypatch):
-    monkeypatch.setattr('builtins.input', lambda: "y")
+    monkeypatch.setattr("builtins.input", lambda: "y")
     # Delete all current officers and salaries
     Salary.query.delete()
     Officer.query.delete()
@@ -427,15 +430,16 @@ def test_csv_new_salary(csvfile, monkeypatch):
         assert float(salary.salary) == 123456.78 or float(salary.salary) == 150000.00
 
 
-def test_bulk_add_officers__success(session, department_with_ranks, csv_path, monkeypatch):
-    monkeypatch.setattr('builtins.input', lambda: "y")
+def test_bulk_add_officers__success(
+    session, department_without_officers, csv_path, monkeypatch
+):
+    monkeypatch.setattr("builtins.input", lambda: "y")
     # generate two officers with different names
-    first_officer = generate_officer()
-    first_officer.department = department_with_ranks
+    first_officer = generate_officer(department_without_officers)
     print(Job.query.all())
-    print(Job.query.filter_by(department=department_with_ranks).all())
+    print(Job.query.filter_by(department=department_without_officers).all())
     job = (
-        Job.query.filter_by(department=department_with_ranks).filter_by(order=1)
+        Job.query.filter_by(department=department_without_officers).filter_by(order=1)
     ).first()
     fo_fn = "Uniquefirst"
     first_officer.first_name = fo_fn
@@ -445,8 +449,7 @@ def test_bulk_add_officers__success(session, department_with_ranks, csv_path, mo
     assignment = Assignment(baseofficer=first_officer, job_id=job.id)
     session.add(assignment)
     session.commit()
-    different_officer = generate_officer()
-    different_officer.department = department_with_ranks
+    different_officer = generate_officer(department_without_officers)
     different_officer.job = job
     do_fn = different_officer.first_name
     do_ln = different_officer.last_name
@@ -455,7 +458,7 @@ def test_bulk_add_officers__success(session, department_with_ranks, csv_path, mo
     session.add(assignment)
     session.commit()
 
-    department_id = department_with_ranks.id
+    department_id = department_without_officers.id
 
     # generate csv to update one existing officer and add one new
 
@@ -526,15 +529,13 @@ def test_bulk_add_officers__duplicate_name(session, department, csv_path):
     # two officers with the same name
     first_name = "James"
     last_name = "Smith"
-    first_officer = generate_officer()
-    first_officer.department = department
+    first_officer = generate_officer(department)
     first_officer.first_name = first_name
     first_officer.last_name = last_name
     session.add(first_officer)
     session.commit()
 
-    different_officer = generate_officer()
-    different_officer.department = department
+    different_officer = generate_officer(department)
     different_officer.first_name = first_name
     different_officer.last_name = last_name
     session.add(different_officer)
@@ -569,12 +570,13 @@ def test_bulk_add_officers__duplicate_name(session, department, csv_path):
     assert isinstance(result.exception, MultipleResultsFound)
 
 
-def test_bulk_add_officers__write_static_null_field(session, department, csv_path, monkeypatch):
-    monkeypatch.setattr('builtins.input', lambda: "y")
+def test_bulk_add_officers__write_static_null_field(
+    session, department, csv_path, monkeypatch
+):
+    monkeypatch.setattr("builtins.input", lambda: "y")
     # start with an officer whose birth_year is missing
-    officer = generate_officer()
+    officer = generate_officer(department)
     officer.birth_year = None
-    officer.department = department
     session.add(officer)
     session.commit()
     fo_uuid = officer.unique_internal_identifier
@@ -616,10 +618,9 @@ def test_bulk_add_officers__write_static_null_field(session, department, csv_pat
 
 def test_bulk_add_officers__write_static_field_no_flag(session, department, csv_path):
     # officer with birth year set
-    officer = generate_officer()
+    officer = generate_officer(department)
     old_birth_year = 1979
     officer.birth_year = old_birth_year
-    officer.department = department
     session.add(officer)
     session.commit()
     fo_uuid = officer.unique_internal_identifier
@@ -661,12 +662,13 @@ def test_bulk_add_officers__write_static_field_no_flag(session, department, csv_
     assert officer.birth_year == old_birth_year
 
 
-def test_bulk_add_officers__write_static_field__flag_set(session, department, csv_path, monkeypatch):
-    monkeypatch.setattr('builtins.input', lambda: "y")
+def test_bulk_add_officers__write_static_field__flag_set(
+    session, department, csv_path, monkeypatch
+):
+    monkeypatch.setattr("builtins.input", lambda: "y")
     # officer with birth year set
-    officer = generate_officer()
+    officer = generate_officer(department)
     officer.birth_year = 1979
-    officer.department = department
     session.add(officer)
     session.commit()
     officer_uuid = officer.unique_internal_identifier
@@ -709,13 +711,14 @@ def test_bulk_add_officers__write_static_field__flag_set(session, department, cs
     assert officer.birth_year == new_birth_year
 
 
-def test_bulk_add_officers__no_create_flag(session, department, csv_path, monkeypatch):
-    monkeypatch.setattr('builtins.input', lambda: "y")
+def test_bulk_add_officers__no_create_flag(
+    session, department_without_officers, csv_path, monkeypatch
+):
+    monkeypatch.setattr("builtins.input", lambda: "y")
     # department with one officer
-    department_id = department.id
-    officer = generate_officer()
+    department_id = department_without_officers.id
+    officer = generate_officer(department_without_officers)
     officer.gender = None
-    officer.department = department
     session.add(officer)
     session.commit()
     officer_uuid = officer.unique_internal_identifier
@@ -762,14 +765,15 @@ def test_bulk_add_officers__no_create_flag(session, department, csv_path, monkey
     assert result.exception is None
 
     # confirm that only one officer is in database and information was updated
+    print(Officer.query.filter_by(department_id=department_id).all())
     officer = Officer.query.filter_by(department_id=department_id).one()
     assert officer.unique_internal_identifier == officer_uuid
     assert officer.gender == officer_gender_updated
 
 
-def test_advanced_csv_import__success(session, department_with_ranks, test_csv_dir):
+def test_advanced_csv_import__success(session, department, test_csv_dir):
     # make sure department name aligns with the csv files
-    assert department_with_ranks.name == "Springfield Police Department"
+    assert department.name == "Springfield Police Department"
 
     # set up existing data
     officer = Officer(
@@ -786,12 +790,16 @@ def test_advanced_csv_import__success(session, department_with_ranks, test_csv_d
         officer_id=officer.id,
         star_no="4567",
         star_date=datetime.date(2020, 1, 1),
-        job_id=department_with_ranks.jobs[0].id,
+        job_id=department.jobs[0].id,
     )
     session.add(assignment)
 
     salary = Salary(
-        id=33001, salary=30000, officer_id=officer.id, year=2018, is_fiscal_year=False,
+        id=33001,
+        salary=30000,
+        officer_id=officer.id,
+        year=2018,
+        is_fiscal_year=False,
     )
     session.add(salary)
 
@@ -813,7 +821,7 @@ def test_advanced_csv_import__success(session, department_with_ranks, test_csv_d
     result = run_command_print_output(
         advanced_csv_import,
         [
-            str(department_with_ranks.name),
+            str(department.name),
             "--officers-csv",
             os.path.join(test_csv_dir, "officers.csv"),
             "--assignments-csv",
@@ -955,10 +963,10 @@ def _create_csv(data, path, csv_file_name):
     return csv_path
 
 
-def test_advanced_csv_import__force_create(session, department_with_ranks, tmp_path):
+def test_advanced_csv_import__force_create(session, department, tmp_path):
     tmp_path = str(tmp_path)
 
-    department_name = department_with_ranks.name
+    department_name = department.name
 
     other_department = Department(name="Other department", short_name="OPD")
     session.add(other_department)
@@ -1034,7 +1042,7 @@ def test_advanced_csv_import__force_create(session, department_with_ranks, tmp_p
     result = run_command_print_output(
         advanced_csv_import,
         [
-            str(department_with_ranks.name),
+            str(department.name),
             "--officers-csv",
             officers_csv,
             "--assignments-csv",
@@ -1075,35 +1083,37 @@ def test_advanced_csv_import__force_create(session, department_with_ranks, tmp_p
     assert cop1.links[0] == link
 
 
-def test_advanced_csv_import__overwrite_assignments(
-    session, department_with_ranks, tmp_path
-):
+def test_advanced_csv_import__overwrite_assignments(session, department, tmp_path):
     tmp_path = str(tmp_path)
 
-    department_name = department_with_ranks.name
+    department_name = department.name
 
     other_department = Department(name="Other department", short_name="OPD")
     session.add(other_department)
 
+    cop1_id = 999001
+    cop2_id = 999002
     officer = Officer(
-        id=99001,
-        department_id=department_with_ranks.id,
+        id=cop1_id,
+        department_id=department.id,
         first_name="Already",
         last_name="InDatabase",
     )
     officer2 = Officer(
-        id=99002,
-        department_id=department_with_ranks.id,
+        id=cop2_id,
+        department_id=department.id,
         first_name="Also",
         last_name="InDatabase",
     )
+    a1_id = 999101
+    a2_id = 999102
     assignment = Assignment(
-        id=123,
+        id=a1_id,
         officer_id=officer.id,
         job_id=Job.query.filter_by(job_title="Police Officer").first().id,
     )
     assignment2 = Assignment(
-        id=124,
+        id=a2_id,
         officer_id=officer2.id,
         job_id=Job.query.filter_by(job_title="Police Officer").first().id,
     )
@@ -1125,17 +1135,19 @@ def test_advanced_csv_import__overwrite_assignments(
 
     officers_csv = _create_csv(officers_data, tmp_path, "officers.csv")
 
+    b1 = "12345"
+    b2 = "999"
     assignments_data = [
         {
-            "officer_id": 99001,
+            "officer_id": cop1_id,
             "job title": "Captain",
-            "badge number": "12345",
+            "badge number": b1,
             "start date": "2020-07-24",
         },
         {
             "officer_id": "#1",
             "job title": "Police Officer",
-            "badge number": "999",
+            "badge number": b2,
             "start date": "2020-07-21",
         },
     ]
@@ -1145,7 +1157,7 @@ def test_advanced_csv_import__overwrite_assignments(
     result = run_command_print_output(
         advanced_csv_import,
         [
-            str(department_with_ranks.name),
+            str(department.name),
             "--officers-csv",
             officers_csv,
             "--assignments-csv",
@@ -1159,24 +1171,22 @@ def test_advanced_csv_import__overwrite_assignments(
     assert result.exit_code == 0
 
     # make sure all the data is imported as expected
-    cop1 = Officer.query.get(99001)
+    cop1 = Officer.query.get(cop1_id)
     assert len(cop1.assignments.all()) == 1
-    assert cop1.assignments[0].star_no == "12345"
+    assert cop1.assignments[0].star_no == b1
 
-    cop2 = Officer.query.get(99002)
+    cop2 = Officer.query.get(cop2_id)
     assert len(cop2.assignments.all()) == 1
-    assert cop2.assignments[0] == Assignment.query.get(124)
+    assert cop2.assignments[0] == Assignment.query.get(a2_id)
 
-    cop3 = Officer.query.filter_by(first_name="Second").first()
+    cop3 = Officer.query.filter_by(first_name="Second", last_name="Test").first()
     assert len(cop3.assignments.all()) == 1
-    assert cop3.assignments[0].star_no == "999"
+    assert cop3.assignments[0].star_no == b2
     assert cop3.assignments[0].job.job_title == "Police Officer"
 
 
-def test_advanced_csv_import__extra_fields_officers(
-    session, department_with_ranks, tmp_path
-):
-    department_name = department_with_ranks.name
+def test_advanced_csv_import__extra_fields_officers(session, department, tmp_path):
+    department_name = department.name
     # create csv with invalid field 'name'
     officers_data = [
         {"id": "", "department_name": department_name, "name": "John Smith"},
@@ -1186,7 +1196,7 @@ def test_advanced_csv_import__extra_fields_officers(
     # run command
     result = run_command_print_output(
         advanced_csv_import,
-        [str(department_with_ranks.name), "--officers-csv", officers_csv],
+        [str(department.name), "--officers-csv", officers_csv],
     )
 
     # expect the command to fail because of unexpected field 'name'
@@ -1196,9 +1206,9 @@ def test_advanced_csv_import__extra_fields_officers(
 
 
 def test_advanced_csv_import__missing_required_field_officers(
-    session, department_with_ranks, tmp_path
+    session, department, tmp_path
 ):
-    department_name = department_with_ranks.name
+    department_name = department.name
     # create csv with missing field 'id'
     officers_data = [
         {
@@ -1212,7 +1222,7 @@ def test_advanced_csv_import__missing_required_field_officers(
     # run command
     result = run_command_print_output(
         advanced_csv_import,
-        [str(department_with_ranks.name), "--officers-csv", officers_csv],
+        [str(department.name), "--officers-csv", officers_csv],
     )
 
     # expect the command to fail because 'id' is missing
@@ -1221,10 +1231,8 @@ def test_advanced_csv_import__missing_required_field_officers(
     assert "id" in str(result.exception)
 
 
-def test_advanced_csv_import__wrong_department(
-    session, department_with_ranks, tmp_path
-):
-    department_name = department_with_ranks.name
+def test_advanced_csv_import__wrong_department(session, department, tmp_path):
+    department_name = department.name
     other_department = Department(name="Other department", short_name="OPD")
     session.add(other_department)
 
@@ -1241,7 +1249,8 @@ def test_advanced_csv_import__wrong_department(
 
     # run command with wrong department name
     result = run_command_print_output(
-        advanced_csv_import, [other_department.name, "--officers-csv", officers_csv],
+        advanced_csv_import,
+        [other_department.name, "--officers-csv", officers_csv],
     )
 
     # expect command to fail because the department name provided to the
@@ -1251,9 +1260,9 @@ def test_advanced_csv_import__wrong_department(
 
 
 def test_advanced_csv_import__update_officer_different_department(
-    session, department_with_ranks, tmp_path
+    session, department, tmp_path
 ):
-    department_name = department_with_ranks.name
+    department_name = department.name
 
     # set up data
     other_department = Department(name="Other department", short_name="OPD")
@@ -1277,7 +1286,7 @@ def test_advanced_csv_import__update_officer_different_department(
     # run command
     result = run_command_print_output(
         advanced_csv_import,
-        [str(department_with_ranks.name), "--officers-csv", officers_csv],
+        [str(department.name), "--officers-csv", officers_csv],
     )
 
     # command fails because the officer is assigned to a different department
@@ -1287,19 +1296,15 @@ def test_advanced_csv_import__update_officer_different_department(
 
 
 def test_advanced_csv_import__unit_other_department(
-    session, department_with_ranks, tmp_path
+    session, department, department_without_officers, tmp_path
 ):
-    department_id = department_with_ranks.id
-
     # set up data
-    officer = generate_officer()
-    officer.department_id = department_id
+    officer = generate_officer(department)
     session.add(officer)
     session.flush()
-    other_department = Department(name="Other department", short_name="OPD")
-    session.add(other_department)
+    session.add(department_without_officers)
     session.flush()
-    unit = Unit(department_id=other_department.id)
+    unit = Unit(department_id=department_without_officers.id)
     session.add(unit)
     session.flush()
 
@@ -1315,7 +1320,7 @@ def test_advanced_csv_import__unit_other_department(
     assignments_csv = _create_csv(assignments_data, tmp_path, "assignments.csv")
     result = run_command_print_output(
         advanced_csv_import,
-        [department_with_ranks.name, "--assignments-csv", assignments_csv],
+        [department.name, "--assignments-csv", assignments_csv],
     )
 
     # command fails because the unit does not belong to the department
@@ -1323,13 +1328,13 @@ def test_advanced_csv_import__unit_other_department(
     assert result.exit_code != 0
 
 
-def test_create_officer_from_row_adds_new_officer_and_normalizes_gender(app, session):
+def test_create_officer_from_row_adds_new_officer_and_normalizes_gender(
+    app, session, department_without_officers
+):
     with app.app_context():
-        department = Department(name="Cityname Police Department", short_name="CNPD")
-        session.add(department)
-        session.commit()
         lookup_officer = Officer.query.filter_by(
-            first_name="NewOfficerFromRow").one_or_none()
+            first_name="NewOfficerFromRow"
+        ).one_or_none()
         assert lookup_officer is None
 
         row = {
@@ -1339,10 +1344,11 @@ def test_create_officer_from_row_adds_new_officer_and_normalizes_gender(app, ses
             "employment_date": "1980-12-01",
             "unique_internal_identifier": "officer-jones-unique-id",
         }
-        create_officer_from_row(row, department.id)
+        create_officer_from_row(row, department_without_officers.id)
 
         lookup_officer = Officer.query.filter_by(
-            first_name="NewOfficerFromRow").one_or_none()
+            first_name="NewOfficerFromRow"
+        ).one_or_none()
 
         # Was an officer created in the database?
         assert lookup_officer is not None
