@@ -1,83 +1,118 @@
-from pytest import raises
-import time
 import datetime
-from OpenOversight.app.models import (Officer, Assignment, Face, Image, Unit,
-                                      User, db, Department, Location, Link,
-                                      LicensePlate, Incident, Salary)
+import time
+
+from pytest import raises
+
+from OpenOversight.app.models import (
+    Assignment,
+    Department,
+    Face,
+    Image,
+    Incident,
+    LicensePlate,
+    Link,
+    Location,
+    Officer,
+    Salary,
+    Unit,
+    User,
+    db,
+)
 
 
 def test_department_repr(mockdata):
     department = Department.query.first()
-    assert department.__repr__() == '<Department ID {}: {}>'.format(department.id, department.name)
+    assert department.__repr__() == "<Department ID {}: {}>".format(
+        department.id, department.name
+    )
 
 
 def test_officer_repr(mockdata):
     officer = Officer.query.first()
     if officer.unique_internal_identifier:
-        assert officer.__repr__() == '<Officer ID {}: {} {} {} {} ({})>'.format(officer.id, officer.first_name, officer.middle_initial, officer.last_name, officer.suffix, officer.unique_internal_identifier)
+        assert officer.__repr__() == "<Officer ID {}: {} {} {} {} ({})>".format(
+            officer.id,
+            officer.first_name,
+            officer.middle_initial,
+            officer.last_name,
+            officer.suffix,
+            officer.unique_internal_identifier,
+        )
     else:
-        assert officer.__repr__() == '<Officer ID {}: {} {} {} {}>'.format(officer.id, officer.first_name, officer.middle_initial, officer.last_name, officer.suffix)
+        assert officer.__repr__() == "<Officer ID {}: {} {} {} {}>".format(
+            officer.id,
+            officer.first_name,
+            officer.middle_initial,
+            officer.last_name,
+            officer.suffix,
+        )
 
 
 def test_assignment_repr(mockdata):
     assignment = Assignment.query.first()
-    assert assignment.__repr__() == '<Assignment: ID {} : {}>'.format(assignment.officer.id, assignment.star_no)
+    assert assignment.__repr__() == "<Assignment: ID {} : {}>".format(
+        assignment.officer.id, assignment.star_no
+    )
 
 
 def test_image_repr(mockdata):
     image = Image.query.first()
-    assert image.__repr__() == '<Image ID {}: {}>'.format(image.id, image.filepath)
+    assert image.__repr__() == "<Image ID {}: {}>".format(image.id, image.filepath)
 
 
 def test_face_repr(mockdata):
     face = Face.query.first()
-    assert face.__repr__() == '<Tag ID {}: {} - {}>'.format(face.id, face.officer_id, face.img_id)
+    assert face.__repr__() == "<Tag ID {}: {} - {}>".format(
+        face.id, face.officer_id, face.img_id
+    )
 
 
 def test_unit_repr(mockdata):
     unit = Unit.query.first()
-    assert unit.__repr__() == 'Unit: {}'.format(unit.descrip)
+    assert unit.__repr__() == "Unit: {}".format(unit.descrip)
 
 
 def test_user_repr(mockdata):
-    user = User(username='bacon')
+    user = User(username="bacon")
     assert user.__repr__() == "<User '{}'>".format(user.username)
 
 
 def test_salary_repr(mockdata):
     salary = Salary.query.first()
-    assert salary.__repr__() == '<Salary: ID {} : {}'.format(salary.officer_id, salary.salary)
+    assert salary.__repr__() == "<Salary: ID {} : {}".format(
+        salary.officer_id, salary.salary
+    )
 
 
 def test_password_not_printed(mockdata):
-    user = User(password='bacon')
+    user = User(password="bacon")
     with raises(AttributeError):
         user.password
 
 
 def test_password_set_success(mockdata):
-    user = User(password='bacon')
+    user = User(password="bacon")
     assert user.password_hash is not None
 
 
 def test_password_verification_success(mockdata):
-    user = User(password='bacon')
-    assert user.verify_password('bacon') is True
+    user = User(password="bacon")
+    assert user.verify_password("bacon") is True
 
 
 def test_password_verification_failure(mockdata):
-    user = User(password='bacon')
-    assert user.verify_password('vegan bacon') is False
+    user = User(password="bacon")
+    assert user.verify_password("vegan bacon") is False
 
 
 def test_password_salting(mockdata):
-    user1 = User(password='bacon')
-    user2 = User(password='bacon')
+    user1 = User(password="bacon")
+    user2 = User(password="bacon")
     assert user1.password_hash != user2.password_hash
 
 
 def test_valid_confirmation_token(mockdata):
-    user = User(password='bacon')
+    user = User(password="bacon")
     db.session.add(user)
     db.session.commit()
     token = user.generate_confirmation_token()
@@ -85,8 +120,8 @@ def test_valid_confirmation_token(mockdata):
 
 
 def test_invalid_confirmation_token(mockdata):
-    user1 = User(password='bacon')
-    user2 = User(password='bacon')
+    user1 = User(password="bacon")
+    user2 = User(password="bacon")
     db.session.add(user1)
     db.session.add(user2)
     db.session.commit()
@@ -95,7 +130,7 @@ def test_invalid_confirmation_token(mockdata):
 
 
 def test_expired_confirmation_token(mockdata):
-    user = User(password='bacon')
+    user = User(password="bacon")
     db.session.add(user)
     db.session.commit()
     token = user.generate_confirmation_token(1)
@@ -104,58 +139,91 @@ def test_expired_confirmation_token(mockdata):
 
 
 def test_valid_reset_token(mockdata):
-    user = User(password='bacon')
+    user = User(password="bacon")
     db.session.add(user)
     db.session.commit()
     token = user.generate_reset_token()
-    assert user.reset_password(token, 'vegan bacon') is True
-    assert user.verify_password('vegan bacon') is True
+    assert user.reset_password(token, "vegan bacon") is True
+    assert user.verify_password("vegan bacon") is True
 
 
 def test_invalid_reset_token(mockdata):
-    user1 = User(password='bacon')
-    user2 = User(password='vegan bacon')
+    user1 = User(password="bacon")
+    user2 = User(password="vegan bacon")
     db.session.add(user1)
     db.session.add(user2)
     db.session.commit()
     token = user1.generate_reset_token()
-    assert user2.reset_password(token, 'tempeh') is False
-    assert user2.verify_password('vegan bacon') is True
+    assert user2.reset_password(token, "tempeh") is False
+    assert user2.verify_password("vegan bacon") is True
+
+
+def test_expired_reset_token(mockdata):
+    user = User(password="bacon")
+    db.session.add(user)
+    db.session.commit()
+    token = user.generate_reset_token(expiration=-1)
+    assert user.reset_password(token, "tempeh") is False
+    assert user.verify_password("bacon") is True
 
 
 def test_valid_email_change_token(mockdata):
-    user = User(email='brian@example.com', password='bacon')
+    user = User(email="brian@example.com", password="bacon")
     db.session.add(user)
     db.session.commit()
-    token = user.generate_email_change_token('lucy@example.org')
+    token = user.generate_email_change_token("lucy@example.org")
     assert user.change_email(token) is True
-    assert user.email == 'lucy@example.org'
+    assert user.email == "lucy@example.org"
+
+
+def test_email_change_token_no_email(mockdata):
+    user = User(email="brian@example.com", password="bacon")
+    db.session.add(user)
+    db.session.commit()
+    token = user.generate_email_change_token(None)
+    assert user.change_email(token) is False
+    assert user.email == "brian@example.com"
 
 
 def test_invalid_email_change_token(mockdata):
-    user1 = User(email='jen@example.com', password='cat')
-    user2 = User(email='freddy@example.com', password='dog')
+    user1 = User(email="jen@example.com", password="cat")
+    user2 = User(email="freddy@example.com", password="dog")
     db.session.add(user1)
     db.session.add(user2)
     db.session.commit()
-    token = user1.generate_email_change_token('mason@example.net')
+    token = user1.generate_email_change_token("mason@example.net")
     assert user2.change_email(token) is False
-    assert user2.email == 'freddy@example.com'
+    assert user2.email == "freddy@example.com"
+
+
+def test_expired_email_change_token(mockdata):
+    user = User(email="jen@example.com", password="cat")
+    db.session.add(user)
+    db.session.commit()
+    token = user.generate_email_change_token("mason@example.net", expiration=-1)
+    assert user.change_email(token) is False
+    assert user.email == "jen@example.com"
 
 
 def test_duplicate_email_change_token(mockdata):
-    user1 = User(email='alice@example.com', password='cat')
-    user2 = User(email='bob@example.org', password='dog')
+    user1 = User(email="alice@example.com", password="cat")
+    user2 = User(email="bob@example.org", password="dog")
     db.session.add(user1)
     db.session.add(user2)
     db.session.commit()
-    token = user2.generate_email_change_token('alice@example.com')
+    token = user2.generate_email_change_token("alice@example.com")
     assert user2.change_email(token) is False
-    assert user2.email == 'bob@example.org'
+    assert user2.email == "bob@example.org"
 
 
 def test_area_coordinator_with_dept_is_valid(mockdata):
-    user1 = User(email='alice@example.com', username='me', password='cat', is_area_coordinator=True, ac_department_id=1)
+    user1 = User(
+        email="alice@example.com",
+        username="me",
+        password="cat",
+        is_area_coordinator=True,
+        ac_department_id=1,
+    )
     db.session.add(user1)
     db.session.commit()
     assert user1.is_area_coordinator is True
@@ -165,24 +233,26 @@ def test_area_coordinator_with_dept_is_valid(mockdata):
 def test_locations_must_have_valid_zip_codes(mockdata):
     with raises(ValueError):
         Location(
-            street_name='Brookford St',
-            cross_street1='Mass Ave',
-            cross_street2='None',
-            city='Cambridge',
-            state='MA',
-            zip_code='543')
+            street_name="Brookford St",
+            cross_street1="Mass Ave",
+            cross_street2="None",
+            city="Cambridge",
+            state="MA",
+            zip_code="543",
+        )
 
 
 def test_locations_can_be_saved_with_valid_zip_codes(mockdata):
-    zip_code = '03456'
-    city = 'Cambridge'
+    zip_code = "03456"
+    city = "Cambridge"
     lo = Location(
-        street_name='Brookford St',
-        cross_street1='Mass Ave',
-        cross_street2='None',
+        street_name="Brookford St",
+        cross_street1="Mass Ave",
+        cross_street2="None",
         city=city,
-        state='MA',
-        zip_code=zip_code)
+        state="MA",
+        zip_code=zip_code,
+    )
     db.session.add(lo)
     db.session.commit()
     saved = Location.query.filter_by(zip_code=zip_code, city=city)
@@ -192,24 +262,26 @@ def test_locations_can_be_saved_with_valid_zip_codes(mockdata):
 def test_locations_must_have_valid_states(mockdata):
     with raises(ValueError):
         Location(
-            street_name='Brookford St',
-            cross_street1='Mass Ave',
-            cross_street2='None',
-            city='Cambridge',
-            state='JK',
-            zip_code='54340')
+            street_name="Brookford St",
+            cross_street1="Mass Ave",
+            cross_street2="None",
+            city="Cambridge",
+            state="JK",
+            zip_code="54340",
+        )
 
 
 def test_locations_can_be_saved_with_valid_states(mockdata):
-    state = 'AZ'
-    city = 'Cambridge'
+    state = "AZ"
+    city = "Cambridge"
     lo = Location(
-        street_name='Brookford St',
-        cross_street1='Mass Ave',
-        cross_street2='None',
+        street_name="Brookford St",
+        cross_street1="Mass Ave",
+        cross_street2="None",
         city=city,
         state=state,
-        zip_code='54340')
+        zip_code="54340",
+    )
 
     db.session.add(lo)
     db.session.commit()
@@ -219,17 +291,16 @@ def test_locations_can_be_saved_with_valid_states(mockdata):
 
 def test_license_plates_must_have_valid_states(mockdata):
     with raises(ValueError):
-        LicensePlate(
-            number='603EEE',
-            state='JK')
+        LicensePlate(number="603EEE", state="JK")
 
 
 def test_license_plates_can_be_saved_with_valid_states(mockdata):
-    state = 'AZ'
-    number = '603RRR'
+    state = "AZ"
+    number = "603RRR"
     lp = LicensePlate(
         number=number,
-        state=state,)
+        state=state,
+    )
 
     db.session.add(lp)
     db.session.commit()
@@ -238,18 +309,14 @@ def test_license_plates_can_be_saved_with_valid_states(mockdata):
 
 
 def test_links_must_have_valid_urls(mockdata):
-    bad_url = 'www.rachel.com'
+    bad_url = "www.rachel.com"
     with raises(ValueError):
-        Link(
-            link_type='video',
-            url=bad_url)
+        Link(link_type="video", url=bad_url)
 
 
 def test_links_can_be_saved_with_valid_urls(mockdata):
-    good_url = 'http://www.rachel.com'
-    li = Link(
-        link_type='video',
-        url=good_url)
+    good_url = "http://www.rachel.com"
+    li = Link(link_type="video", url=good_url)
     db.session.add(li)
     db.session.commit()
     saved = Link.query.filter_by(url=good_url).first()
@@ -258,12 +325,14 @@ def test_links_can_be_saved_with_valid_urls(mockdata):
 
 def test_incident_m2m_officers(mockdata):
     incident = Incident.query.first()
-    officer = Officer(first_name='Test',
-                      last_name='McTesterson',
-                      middle_initial='T',
-                      race='WHITE',
-                      gender='M',
-                      birth_year=1990)
+    officer = Officer(
+        first_name="Test",
+        last_name="McTesterson",
+        middle_initial="T",
+        race="WHITE",
+        gender="M",
+        birth_year=1990,
+    )
     incident.officers.append(officer)
     db.session.add(incident)
     db.session.add(officer)
@@ -274,9 +343,7 @@ def test_incident_m2m_officers(mockdata):
 
 def test_incident_m2m_links(mockdata):
     incident = Incident.query.first()
-    link = Link(
-        link_type='video',
-        url='http://www.lulz.com')
+    link = Link(link_type="video", url="http://www.lulz.com")
     incident.links.append(link)
     db.session.add(incident)
     db.session.add(link)
@@ -288,8 +355,9 @@ def test_incident_m2m_links(mockdata):
 def test_incident_m2m_license_plates(mockdata):
     incident = Incident.query.first()
     license_plate = LicensePlate(
-        number='W23F43',
-        state='DC',)
+        number="W23F43",
+        state="DC",
+    )
     incident.license_plates.append(license_plate)
     db.session.add(incident)
     db.session.add(license_plate)
@@ -300,12 +368,15 @@ def test_incident_m2m_license_plates(mockdata):
 
 def test_images_added_with_user_id(mockdata):
     user_id = 1
-    new_image = Image(filepath="http://www.example.com", hash_img="1234",
-                      is_tagged=False,
-                      date_image_inserted=datetime.datetime.now(),
-                      department_id=1,
-                      date_image_taken=datetime.datetime.now(),
-                      user_id=user_id)
+    new_image = Image(
+        filepath="http://www.example.com",
+        hash_img="1234",
+        is_tagged=False,
+        date_image_inserted=datetime.datetime.now(),
+        department_id=1,
+        date_image_taken=datetime.datetime.now(),
+        user_id=user_id,
+    )
     db.session.add(new_image)
     db.session.commit()
     saved = Image.query.filter_by(user_id=user_id).first()
