@@ -2,6 +2,8 @@
 from OpenOversight.app import create_app
 from OpenOversight.app.models import db
 
-app = create_app('development')
-db.app = app
-db.create_all()
+
+app = create_app("development")
+with app.app_context():
+    db.app = app
+    db.create_all()
